% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{AAPL}
\alias{AAPL}
\title{Apple Inc. Stock Daily Price Data}
\format{
An xts time series object with 5 columns:
\itemize{
\item Open: Opening price
\item High: Intraday high price
\item Low: Intraday low price
\item Close: Closing price
\item Volume: Daily trading volume
}
}
\source{
Yahoo Finance
}
\usage{
data(AAPL)
}
\description{
Historical daily OHLCV time series data for Apple Inc. ticker AAPL.
}
\keyword{datasets}
