% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_momentum.R
\name{add_momentum}
\alias{add_momentum}
\title{Add Momentum (Panel Data)}
\usage{
add_momentum(
  mkt_data,
  n = 1,
  na.pad = TRUE,
  append = TRUE,
  output = c("tibble", "data.frame")
)
}
\arguments{
\item{mkt_data}{Long-format panel with date, code, close.}

\item{n}{Number of periods for momentum. Default 1.}

\item{na.pad}{Pad leading NAs. Default TRUE.}

\item{append}{Logical.}

\item{output}{"tibble" or "data.frame".}
}
\value{
Data frame with added column MOMENTUM_\if{html}{\out{<n>}}.
}
\description{
Add Momentum (Panel Data)
}
