% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_ROC.R
\name{add_ROC}
\alias{add_ROC}
\title{Add Rate of Change (Panel Data)}
\usage{
add_ROC(
  mkt_data,
  n = 1,
  type = c("discrete", "continuous"),
  na.pad = TRUE,
  append = TRUE,
  output = c("tibble", "data.frame")
)
}
\arguments{
\item{mkt_data}{Panel with date, code, close.}

\item{n}{Lag periods. Default 1.}

\item{type}{"continuous" (log diff) or "discrete" (ratio). Default "discrete".}

\item{na.pad}{Pad leading NAs. Default TRUE.}

\item{append}{Logical.}

\item{output}{"tibble" or "data.frame".}
}
\value{
Data frame with column ROC_\if{html}{\out{<n>}}.
}
\description{
Add Rate of Change (Panel Data)
}
