% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EMA.R
\name{EMA}
\alias{EMA}
\title{Exponential Moving Average (EMA)}
\usage{
EMA(x, n = 10, wilder = FALSE, ratio = NULL)
}
\arguments{
\item{x}{Price series that is coercible to xts or matrix.}

\item{n}{Number of periods to average over. Must be between 1 and \code{nrow(x)}.}

\item{wilder}{logical; if \code{TRUE}, use Welles Wilder's EMA formula (\code{1/n}).}

\item{ratio}{A smoothing/decay ratio (overrides \code{wilder}).}
}
\value{
An object of the same class as \code{x} containing the EMA values.
}
\description{
Calculate an exponentially-weighted mean, giving more weight to recent observations.
}
\note{
\code{wilder=FALSE} uses \code{2/(n+1)}; \code{wilder=TRUE} uses \code{1/n}.
}
\examples{
data(TSLA)
ema_20 <- EMA(TSLA[, "Close"], 20)
head(ema_20)
}
\keyword{ts}
