% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eTTR-package.R
\docType{package}
\name{eTTR}
\alias{eTTR}
\alias{eTTR-package}
\title{Functions to create Technical Trading Rules (TTR)}
\description{
This package contains many of the most popular technical analysis functions,
as well as functions to retrieve U.S. stock symbols, and data from Yahoo
Finance.
}
\details{
Users will probably be most interested in the following functions:\cr
\code{\link[TTR]{ADX}}\cr \code{\link[TTR]{BBands}}\cr \code{\link[TTR]{MACD}}\cr
\code{\link[TTR]{RSI}}\cr \code{\link[TTR]{stoch}}\cr \code{\link[TTR]{VWAP}}\cr

Global import functions from external packages, effective for all package scripts
}
\references{
The following sites were used to code/document this package:\cr
\url{https://www.fmlabs.com/reference/default.htm}\cr
\url{https://www.metastock.com/Customer/Resources/TAAZ/}\cr
\url{https://www.linnsoft.com/indicators}\cr
\url{https://school.stockcharts.com/doku.php?id=technical_indicators}\cr
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/dengyishuo/eTTR}
  \item Report bugs at \url{https://github.com/dengyishuo/eTTR/issues}
}

}
\author{
DengYishuo
}
\keyword{internal}
\keyword{package}
