% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EVWMA.R
\name{EVWMA}
\alias{EVWMA}
\title{Elastic Volume-Weighted Moving Average (EVWMA)}
\usage{
EVWMA(price, volume, n = 10)
}
\arguments{
\item{price}{Price series that is coercible to xts or matrix.}

\item{volume}{Volume series (must match length of \code{price} or be constant).}

\item{n}{Number of periods to average over.}
}
\value{
An object of the same class as \code{price} containing the EVWMA values.
}
\description{
Calculate a volume-weighted moving average that adjusts to trading volume.
}
\examples{
data(ttrc)
evwma_20 <- EVWMA(ttrc[, "Close"], ttrc[, "Volume"], 20)
head(evwma_20)
}
\keyword{ts}
