% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth.R
\name{growth}
\alias{growth}
\title{Calculate Investment Growth}
\usage{
growth(price, signals, ...)
}
\arguments{
\item{price}{Price series that is coercible to xts or matrix.}

\item{signals}{Signals to use (defaults to vector of ones). Use '0' for no position,
'1' for long position, and '-1' for short position.}

\item{...}{Further arguments to be passed to ROC for return calculation.}
}
\value{
A vector of the cumulative growth of the investment over time.
}
\description{
This function computes the growth of an investment based on given prices
and trading signals. It assumes an initial investment of $1 and applies
the provided signals to determine positions (long, short, or none).
}
\note{
You can specify the number of periods and type of compounding
via the '...' argument when calculating returns.
}
\examples{
prices <- c(100, 105, 103, 107, 110)
signals <- c(1, 1, -1, 1, 1)
growth(prices, signals)

}
\author{
DengYishuo
}
\keyword{ts}
