% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HMA.R
\name{HMA}
\alias{HMA}
\title{Hull Moving Average (HMA)}
\usage{
HMA(x, n = 20)
}
\arguments{
\item{x}{Price series that is coercible to xts or matrix.}

\item{n}{Number of periods to average over.}
}
\value{
An object of the same class as \code{x} containing the HMA values.
}
\description{
Calculate a highly responsive moving average with reduced lag.
}
\examples{
data(ttrc)
hma_20 <- HMA(ttrc[, "Close"], 20)
head(hma_20)
}
\keyword{ts}
