% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lags.R
\name{lags}
\alias{lags}
\title{Lags Calculation}
\usage{
lags(x, n = 1)
}
\arguments{
\item{x}{Object that is coercible to xts or matrix.}

\item{n}{Number of periods to lag (default is 1).}

\item{...}{Further arguments to be passed from or to other methods.}
}
\value{
A matrix of lagged values with columns labeled by original names and lag periods.
}
\description{
Computes lagged values of a time series or matrix.
}
\details{
Calculate Lags of a Time Series

This function computes lagged values of a given time series or matrix.
It is useful for creating lagged predictors in time series analysis.
}
\note{
This function is deprecated in favor of xts::lag.xts and quantmod::Lag.
}
\examples{
data <- 1:10
lags(data, 2)

}
\author{
DengYishuo
}
\keyword{ts}
