% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/naCheck.R
\name{naCheck}
\alias{naCheck}
\title{NA Check}
\usage{
naCheck(x, n = 0)
}
\arguments{
\item{x}{Object that is coercible to xts or matrix.}

\item{n}{Starting position offset (default is 0).}
}
\value{
An invisible list containing:
\item{NAs}{Number of leading NA values.}
\item{nonNA}{Indices of non-NA values.}
\item{beg}{Starting index after accounting for NAs and offset.}
}
\description{
Validates that NA values are only present at the start of a series.
}
\examples{
data <- c(NA, NA, 3, 4, 5)
naCheck(data)

}
\author{
DengYishuo
}
\keyword{ts}
