% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OBV.R
\name{OBV}
\alias{OBV}
\title{On Balance Volume (OBV)}
\usage{
OBV(price, volume)
}
\arguments{
\item{price}{Price series that is coercible to xts or matrix.}

\item{volume}{Volume series that is coercible to xts or matrix,
corresponding to the price object.}
}
\value{
A object of the same class as \code{price} and \code{volume} or a
        vector (if \code{try.xts} fails) containing the OBV values.
}
\description{
On Balance Volume (OBV) is a momentum-based technical indicator that measures
the cumulative flow of volume to determine the strength of a price trend.
It uses volume flow to predict changes in stock prices by adding volume on
up days and subtracting volume on down days.
}
\details{
This function is part of the eTTR (Enhanced Technical Trading Rules) package.
}
\note{
OBV is typically used to confirm price trends or identify potential
      divergences between volume flow and price movements.
}
\section{Copyright}{

eTTR: Enhanced Technical Trading Rules

Copyright (C) 2025-2030  DengYishuo

This program is free software: you can redistribute it and/or modify
it under the terms of the GNU General Public License as published by
the Free Software Foundation, either version 2 of the License, or
(at your option) any later version.

This program is distributed in the hope that it will be useful,
but WITHOUT ANY WARRANTY; without even the implied warranty of
MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
GNU General Public License for more details.

You should have received a copy of the GNU General Public License
along with this program.  If not, see <http://www.gnu.org/licenses/>.
}

\examples{
data(TSLA)
obv <- OBV(TSLA[, "Close"], ttrc[, "Volume"])

}
\references{
- https://www.fmlabs.com/reference/OBV.htm
- https://www.metastock.com/Customer/Resources/TAAZ/?p=82
- https://www.linnsoft.com/techind/balance-open-interest
- https://school.stockcharts.com/doku.php?id=technical_indicators:on_balance_volume_obv
}
\seealso{
See \code{\link{chaikinAD}} for a similar volume-based indicator.
}
\author{
DengYishuo
}
\keyword{indicator}
\keyword{momentum}
\keyword{ts}
\keyword{volume}
