% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runRegression.R
\name{print.runRegression}
\alias{print.runRegression}
\title{Print Method for runRegression Objects}
\usage{
\method{print}{runRegression}(x, ...)
}
\arguments{
\item{x}{A runRegression object.}

\item{...}{Additional arguments passed to print.}
}
\value{
The input object is returned invisibly.
}
\description{
Print Method for runRegression Objects
}
