% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SMA.R
\name{SMA}
\alias{SMA}
\title{Simple Moving Average (SMA)}
\usage{
SMA(x, n = 10)
}
\arguments{
\item{x}{Price series that is coercible to xts or matrix.}

\item{n}{Number of periods to average over. Must be between 1 and \code{nrow(x)}.}
}
\value{
An object of the same class as \code{x} containing the SMA values.
}
\description{
Calculate the arithmetic mean of the series over the past \code{n} observations.
}
\examples{
data(TSLA)
sma_20 <- SMA(TSLA[, "Close"], 20)
head(sma_20)
}
\keyword{ts}
