% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{TSLA}
\alias{TSLA}
\title{Tesla Inc. Stock Daily Price Data}
\format{
An xts time series object with 5 columns:
\itemize{
\item Open: Opening price
\item High: Intraday high price
\item Low: Intraday low price
\item Close: Closing price
\item Volume: Daily trading volume
}
}
\source{
Yahoo Finance
}
\usage{
data(TSLA)
}
\description{
Historical daily OHLCV time series data for Tesla Inc. ticker TSLA.
}
\keyword{datasets}
