% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VWMA.R
\name{VWAP}
\alias{VWAP}
\title{Volume-Weighted Average Price (VWAP)}
\usage{
VWAP(price, volume, n = 10)
}
\arguments{
\item{price}{Price series that is coercible to xts or matrix.}

\item{volume}{Volume series that matches the length of \code{price}.}

\item{n}{Number of periods to average over.}
}
\value{
An object of the same class as \code{price} containing the VWAP values.
}
\description{
Calculate the volume-weighted moving average price.
}
\examples{
data(ttrc)
vwap_20 <- VWAP(ttrc[, "Close"], ttrc[, "Volume"], 20)
head(vwap_20)
}
\keyword{ts}
