% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZLEMA.R
\name{ZLEMA}
\alias{ZLEMA}
\title{Zero-Lag Exponential Moving Average (ZLEMA)}
\usage{
ZLEMA(x, n = 10, ratio = NULL)
}
\arguments{
\item{x}{Price series that is coercible to xts or matrix.}

\item{n}{Number of periods to average over.}

\item{ratio}{A smoothing/decay ratio.}
}
\value{
An object of the same class as \code{x} containing the ZLEMA values.
}
\description{
Calculate an exponential moving average with reduced lag.
}
\examples{
data(ttrc)
zlema_20 <- ZLEMA(ttrc[, "Close"], 20)
head(zlema_20)
}
\keyword{ts}
