% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eTTR-package.R
\docType{data}
\name{AAPL}
\alias{AAPL}
\title{Apple Inc. Stock Price Data}
\format{
A data frame with 5 columns:
\describe{
\item{Open}{Opening price}
\item{High}{Highest price}
\item{Low}{Lowest price}
\item{Close}{Closing price}
\item{Volume}{Trading volume}
}
}
\source{
Yahoo Finance
}
\description{
Daily stock prices for Apple Inc. (AAPL) from 2010 to present.
}
\examples{
data(AAPL)
plot(tail(AAPL[, "Close"], 100), type = "l")
}
\keyword{AAPL}
