% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_ALMA.R
\name{add_ALMA}
\alias{add_ALMA}
\title{Calculate Arnaud Legoux Moving Average (ALMA)}
\usage{
add_ALMA(OHLCV, n = 9, offset = 0.85, sigma = 6, append = FALSE)
}
\arguments{
\item{OHLCV}{Object that is coercible to xts or matrix and contains Open - High - Low - Close - Volume data.
The function will extract the closing price from this object for ALMA calculation.}

\item{n}{Number of periods to average over. This parameter determines the window size of the moving average.
A larger \code{n} will result in a smoother ALMA, but it may be less responsive to recent price changes.}

\item{offset}{Percentile for weight distribution center (0 - 1). A higher \code{offset} value emphasizes
more recent prices in the moving average calculation. For example, an \code{offset} close to 1 will
give more weight to the most recent data points.}

\item{sigma}{Standard deviation of the Gaussian distribution. A lower \code{sigma} reduces the smoothing
effect, making the ALMA more sensitive to short - term price fluctuations.}

\item{append}{A logical value. If \code{TRUE}, the calculated ALMA values will be appended to the \code{OHLCV} input data,
ensuring proper alignment of time - series data. If \code{FALSE}, only the calculated
ALMA values will be returned. Defaults to \code{FALSE}.}
}
\value{
If \code{append} is \code{FALSE}, an object of the same class as \code{OHLCV} (or a vector if \code{try.xts} fails)
containing the ALMA values.
If \code{append} is \code{TRUE}, an object of the same class as \code{OHLCV} with the calculated ALMA values appended,
maintaining the integrity of the time - series alignment.
}
\description{
Calculate a Gaussian - weighted moving average with reduced lag. The Arnaud Legoux Moving Average (ALMA)
uses a Gaussian - shaped weight distribution to calculate the moving average. This results in a more
responsive moving average compared to traditional moving averages, as it can be adjusted to emphasize
recent prices more effectively.
}
\note{
Higher \code{offset} emphasizes recent prices; lower \code{sigma} reduces smoothing.
It's important to note that the choice of \code{offset} and \code{sigma} can significantly impact the
behavior of the ALMA. Traders should carefully consider these parameters based on the characteristics
of the price series and their trading strategies.
}
\examples{
\dontrun{
data(TSLA)
# Using default parameters without appending
alma_result1 <- add_ALMA(TSLA)

# Using default parameters and appending
alma_result2 <- add_ALMA(TSLA, append = TRUE)

# Changing n and without appending
alma_result3 <- add_ALMA(TSLA, n = 12)

# Changing n and appending
alma_result4 <- add_ALMA(TSLA, n = 12, append = TRUE)

# Changing offset and without appending
alma_result5 <- add_ALMA(TSLA, offset = 0.9)

# Changing sigma and without appending
alma_result6 <- add_ALMA(TSLA, sigma = 4)
}
}
\keyword{ts}
