% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_aroon.R
\name{add_aroon}
\alias{add_aroon}
\alias{add_Aroon}
\title{Aroon}
\usage{
add_aroon(OHLCV, n = 20, append = FALSE)
}
\arguments{
\item{OHLCV}{Object that is coercible to xts or matrix and contains Open - High - Low - Close - Volume data.}

\item{n}{Number of periods to use in the calculation. Defaults to 20.}

\item{append}{A logical value. If \code{TRUE}, the calculated result columns
(aroonUp, aroonDn, oscillator) will be appended to the \code{OHLCV} input data, ensuring
proper alignment of time - series data. If \code{FALSE}, only the calculated
result data will be returned. Defaults to \code{FALSE}.}
}
\value{
If \code{append = FALSE}, an object of the same class as \code{OHLCV}
(or a matrix if \code{try.xts} fails) containing the columns:
\describe{
\item{aroonUp}{The Aroon up indicator.}
\item{aroonDn}{The Aroon down indicator.}
\item{oscillator}{The Aroon oscillator (\code{aroonUp - aroonDn}).}
}
If \code{append = TRUE}, an object of the same class as \code{OHLCV} with the
calculated columns appended, maintaining the integrity of the time - series
alignment.
}
\description{
The Aroon indicator attempts to identify starting trends. The indicator
consists of up and down lines, which measure how long it has been since the
highest high / lowest low has occurred in the last \code{n} periods. The Aroon
indicator is developed by Tushar Chande in 1995.
}
\details{
Aroon up (down) is the elapsed time, expressed as a percentage, between today
and the highest (lowest) price in the last \code{n} periods. If today's
price is a new high (low) Aroon up (down) will be 100. Each subsequent period
without another new high (low) causes Aroon up (down) to decrease by \eqn{(1/
n*100}.
}
\note{
If High - Low prices are given, the function calculates the max/min using
the high/low prices. Otherwise the function calculates the max/min of the
single series.
Up (down) trends are indicated when the aroonUp(Dn) is between 70 and 100.
Strong trends are indicated when when the aroonUp(Dn) is above 70 while the
aroonDn(Up) is below 30. Also, crossovers may be useful.
}
\examples{
\dontrun{
# Load required data, assume TSLA is data in the appropriate format
data(TSLA)

# Use default parameters
aroon_result1 <- add_aroon(TSLA)

# Modify the n parameter
aroon_result2 <- add_aroon(TSLA, n = 30)

# Set append to TRUE
aroon_result3 <- add_aroon(TSLA, append = TRUE)

# Combine modifying the n parameter and setting append to TRUE
aroon_result4 <- add_aroon(TSLA, n = 30, append = TRUE)
}
}
\references{
The following site(s) were used to code/document this
indicator:\cr \url{https://www.fmlabs.com/reference/Aroon.htm}\cr
\url{https://www.fmlabs.com/reference/AroonOscillator.htm}\cr
\url{https://www.linnsoft.com/techind/aroon - arn}\cr
\url{https://school.stockcharts.com/doku.php?id=technical_indicators:aroon}\cr
}
\seealso{
See \code{\link{CCI}}, \code{\link{ADX}}, \code{\link{TDI}},
\code{\link{VHF}}, \code{\link{GMMA}} for other indicators that measure trend
direction/strength.
}
\author{
DengYishuo
}
\keyword{ts}
