% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_ATR.R
\name{add_ATR}
\alias{add_ATR}
\title{Calculate Average True Range (ATR)}
\usage{
add_ATR(OHLCV, n = 14, maType, append = FALSE, ...)
}
\arguments{
\item{OHLCV}{An OHLCV (Open - High - Low - Close - Volume) price series, typically an xts object.}

\item{n}{The number of periods to use for the moving average calculation. Defaults to 14.}

\item{maType}{The type of moving average to use. Defaults to "EMA" (Wilder's EMA).}

\item{append}{A logical value. If \code{TRUE}, the calculated result columns
("tr", "atr", "trueHigh", "trueLow") will be appended to the \code{OHLCV} input data, ensuring
proper alignment of time - series data. If \code{FALSE}, only the calculated
result data will be returned. Defaults to \code{FALSE}.}

\item{...}{Additional arguments to be passed to the moving average function.}
}
\value{
If \code{append = FALSE}, an xts object containing columns for the true range ("tr"), the ATR ("atr"),
and the true high and low values ("trueHigh", "trueLow").
If \code{append = TRUE}, an xts object of the same class as \code{OHLCV} with the
calculated columns appended, maintaining the integrity of the time - series
alignment.
}
\description{
This function computes the Average True Range (ATR), a technical analysis indicator
used to measure market volatility. It calculates the true range (TR) and applies a moving
average (defaulting to Wilder's EMA) to smooth the results.
}
\details{
The Average True Range (ATR) is a measure of volatility introduced by J. Welles Wilder.
It considers the entire range of price movement in a given period, including gaps between
sessions. The default method uses Wilder's EMA, which applies a smoothing factor of 1/n.
}
\examples{
\dontrun{
# Calculate ATR for a stock's OHLCV data
data(TSLA)
atr_values <- add_ATR(TSLA, n = 14)

# Using default parameters and appending results
atr_append1 <- add_ATR(TSLA, append = TRUE)

# Modifying n and appending results
atr_append2 <- add_ATR(TSLA, n = 20, append = TRUE)
}
}
\references{
Wilder, J. Welles. "New Concepts in Technical Trading Systems." 1978.
}
\author{
DengYishuo
}
