% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_CCI.R
\name{add_CCI}
\alias{add_CCI}
\title{Calculate Commodity Channel Index}
\usage{
add_CCI(OHLCV, n = 20, maType, c = 0.015, append = FALSE, ...)
}
\arguments{
\item{OHLCV}{Object that is coercible to xts or matrix and contains
Open - High - Low - Close - Volume prices.}

\item{n}{Number of periods for moving average. Defaults to 20.}

\item{maType}{A function or a string naming the function to be called.}

\item{c}{Constant to apply to the mean deviation. Defaults to 0.015.}

\item{append}{A logical value. If \code{TRUE}, the calculated CCI values will be appended to the \code{OHLCV} input data, ensuring
proper alignment of time - series data. If \code{FALSE}, only the calculated
CCI values will be returned. Defaults to \code{FALSE}.}

\item{...}{Other arguments to be passed to the \code{maType} function.}
}
\value{
If \code{append = FALSE}, an object of the same class as \code{OHLCV}
or a vector (if \code{try.xts} fails) containing the CCI values.
If \code{append = TRUE}, an object of the same class as \code{OHLCV} with the
calculated CCI values appended, maintaining the integrity of the time - series
alignment.
}
\description{
The Commodity Channel Index (CCI) attempts to identify starting and ending
trends.
}
\details{
CCI relates the current price and the average of price over \code{n} periods.
The CCI usually falls in a channel of - 100 to 100. A basic CCI trading system
is: Buy (sell) if CCI rises above 100 (falls below - 100) and sell (buy) when
it falls below 100 (rises above - 100).
CCI is usually calculated using the typical price, but if a univariate series
(e.g. Close, Weighted Close, Median Price, etc.) is provided, it will be used
instead.
}
\note{
If \code{OHLCV} contains High - Low - Close prices, then typical price will be
calculated.  If \code{OHLCV} is a univariate series, then those values will be used
instead of the typical price.
}
\examples{
\dontrun{
data(TSLA)
# Using default parameters without appending
cci_result1 <- add_CCI(TSLA)

# Modifying n and without appending
cci_result2 <- add_CCI(TSLA, n = 30)

# Using default parameters and appending
cci_result3 <- add_CCI(TSLA, append = TRUE)

# Modifying n and appending
cci_result4 <- add_CCI(TSLA, n = 30, append = TRUE)
}
}
\references{
The following site(s) were used to code/document this
indicator:\cr \url{https://www.fmlabs.com/reference/CCI.htm}\cr
\url{https://www.metastock.com/Customer/Resources/TAAZ/?p=42}\cr
\url{https://www.linnsoft.com/techind/cci - commodity - channel - index}\cr
\url{https://school.stockcharts.com/doku.php?id=technical_indicators:commodity_channel_index_cci}\cr
}
\seealso{
See \code{\link{EMA}}, \code{\link{SMA}}, etc. for moving average
options; and note Warning section.  See \code{\link{aroon}},
\code{\link{ADX}}, \code{\link{TDI}}, \code{\link{VHF}}, \code{\link{GMMA}}
for other indicators that measure trend direction/strength.
}
\author{
DengYishuo
}
\keyword{ts}
