% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_chaikinAD.R
\name{add_chaikinAD}
\alias{add_chaikinAD}
\title{Calculate Chaikin Accumulation / Distribution}
\usage{
add_chaikinAD(OHLCV, append = FALSE)
}
\arguments{
\item{OHLCV}{Object that is coercible to xts or matrix and contains
Open - High - Low - Close - Volume prices.}

\item{append}{A logical value. If \code{TRUE}, the calculated accumulation / distribution
values will be appended to the \code{OHLCV} input data, ensuring
proper alignment of time - series data. If \code{FALSE}, only the calculated
accumulation / distribution values will be returned. Defaults to \code{FALSE}.}
}
\value{
If \code{append = FALSE}, an object of the same class as \code{OHLCV}
or a vector (if \code{try.xts} fails) containing the accumulation / distribution
values.
If \code{append = TRUE}, an object of the same class as \code{OHLCV} with the
calculated accumulation / distribution values appended, maintaining the integrity of the time - series
alignment.
}
\description{
The Chaikin Accumulation / Distribution (AD) line is a measure of the money
flowing into or out of a security.  It is similar to On Balance Volume (OBV).
Developed by Marc Chaikin.

The AD line is similar to OBV; the difference is that OBV sums volume
multiplied by +/- 1 if the close is higher/lower than the previous close,
while the AD line multiplies volume by the close location value (CLV).
}
\note{
The Accumulation/Distribution Line is interpreted by looking for a
divergence in the direction of the indicator relative to price.
}
\examples{
\dontrun{
data(TSLA)
# Using default parameters without appending
ad_result1 <- add_chaikinAD(TSLA)

# Using default parameters and appending
ad_result2 <- add_chaikinAD(TSLA, append = TRUE)
}
}
\references{
The following site(s) were used to code/document this
indicator:\cr \url{https://www.fmlabs.com/reference/AccumDist.htm}\cr
\url{https://www.metastock.com/Customer/Resources/TAAZ/?p=27}\cr
\url{https://www.linnsoft.com/techind/accumulation - distribution}\cr
\url{https://school.stockcharts.com/doku.php?id=technical_indicators:accumulation_distribution_line}\cr
}
\seealso{
See \code{\link{OBV}}, and \code{\link{CLV}}.
}
\author{
DengYishuo
}
\keyword{ts}
