% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_CLV.R
\name{add_CLV}
\alias{add_CLV}
\title{Calculate Close Location Value}
\usage{
add_CLV(OHLCV, append = FALSE)
}
\arguments{
\item{OHLCV}{Object that is coercible to xts or matrix and contains Open - High - Low - Close - Volume prices.}

\item{append}{A logical value. If \code{TRUE}, the calculated Close Location Values
will be appended to the \code{OHLCV} input data, ensuring
proper alignment of time - series data. If \code{FALSE}, only the calculated
Close Location Values will be returned. Defaults to \code{FALSE}.}
}
\value{
If \code{append = FALSE}, an object of the same class as \code{OHLCV}
or a vector (if \code{try.xts} fails) containing the Close Location Values of a
High - Low - Close price series.
If \code{append = TRUE}, an object of the same class as \code{OHLCV} with the
calculated Close Location Values appended, maintaining the integrity of the time - series
alignment.
}
\description{
The Close Location Value (CLV) relates the day's close to its trading range.
The CLV will fall in a range of -1 to +1.  If the CLV is +/-1, the close is
at the high/low; if the CLV is 0, the close is directly between the high and
low.
}
\examples{
\dontrun{
data(TSLA)
# Using default parameters without appending
clv_result1 <- add_CLV(TSLA)

# Using default parameters and appending
clv_result2 <- add_CLV(TSLA, append = TRUE)
}
}
\references{
The following site(s) were used to code/document this
indicator:\cr
\url{https://school.stockcharts.com/doku.php?id=technical_indicators:accumulation_distribution_line}\cr
}
\seealso{
See \code{\link{chaikinAD}}, which uses CLV.
}
\author{
DengYishuo
}
\keyword{ts}
