% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_CMF.R
\name{add_CMF}
\alias{add_CMF}
\title{Calculate Chaikin Money Flow}
\usage{
add_CMF(OHLCV, n = 20, append = FALSE)
}
\arguments{
\item{OHLCV}{Object that is coercible to xts or matrix and contains Open - High - Low - Close - Volume prices.}

\item{n}{Number of periods to use. Defaults to 20.}

\item{append}{A logical value. If \code{TRUE}, the calculated Chaikin Money Flow
values will be appended to the \code{OHLCV} input data, ensuring
proper alignment of time - series data. If \code{FALSE}, only the calculated
Chaikin Money Flow values will be returned. Defaults to \code{FALSE}.}
}
\value{
If \code{append = FALSE}, an object of the same class as \code{OHLCV}
or a vector (if \code{try.xts} fails) containing the Chaikin Money Flow values.
If \code{append = TRUE}, an object of the same class as \code{OHLCV} with the
calculated Chaikin Money Flow values appended, maintaining the integrity of the time - series
alignment.
}
\description{
Chaikin Money Flow compares total volume over the last \code{n} time periods
to total volume times the Close Location Value (CLV) over the last \code{n}
time periods. Developed by Marc Chaikin.
Chaikin Money Flow is calculated by taking dividing the sum of the Chaikin
Accumulation / Distribution line over the past \code{n} periods by the sum of
volume over the past \code{n} periods.
}
\note{
When Chaikin Money Flow is above/below +/- 0.25 it is a bullish/bearish
signal. If Chaikin Money Flow remains below zero while the price is rising,
it indicates a probable reversal.
}
\examples{
\dontrun{
data(TSLA)
# Using default parameters without appending
cmf_result1 <- add_CMF(TSLA)

# Modifying n and without appending
cmf_result2 <- add_CMF(TSLA, n = 30)

# Using default parameters and appending
cmf_result3 <- add_CMF(TSLA, append = TRUE)

# Modifying n and appending
cmf_result4 <- add_CMF(TSLA, n = 30, append = TRUE)
}
}
\references{
The following site(s) were used to code/document this
indicator:\cr \url{https://www.fmlabs.com/reference/ChaikinMoneyFlow.htm}\cr
\url{https://www.linnsoft.com/techind/chaikin - money - flow - cmf}\cr
\url{https://school.stockcharts.com/doku.php?id=technical_indicators:chaikin_money_flow_cmf}\cr
}
\seealso{
See \code{\link{CLV}}, and \code{\link{chaikinAD}}.
}
\author{
DengYishuo
}
\keyword{ts}
