% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_CMO.R
\name{add_CMO}
\alias{add_CMO}
\title{Calculate Chande Momentum Oscillator}
\usage{
add_CMO(OHLCV, n = 14, append = FALSE)
}
\arguments{
\item{OHLCV}{Price, volume, etc. series that is coercible to xts or matrix, assumed to contain Open - High - Low - Close - Volume data.}

\item{n}{Number of periods to use. Defaults to 14.}

\item{append}{A logical value. If \code{TRUE}, the calculated Chande Momentum Oscillator
values will be appended to the \code{OHLCV} input data, ensuring
proper alignment of time - series data. If \code{FALSE}, only the calculated
Chande Momentum Oscillator values will be returned. Defaults to \code{FALSE}.}
}
\value{
If \code{append = FALSE}, an object of the same class as \code{OHLCV}
or a vector (if \code{try.xts} fails) containing Chande Momentum Oscillator values.
If \code{append = TRUE}, an object of the same class as \code{OHLCV} with the
calculated Chande Momentum Oscillator values appended, maintaining the integrity of the time - series
alignment.
}
\description{
The Chande Momentum Oscillator (CMO) is a modified RSI. Developed by Tushar
S. Chande.
The CMO divides the total movement by the net movement ((up - down) / (up + down)),
where RSI divides the upward movement by the net movement (up / (up + down)).
}
\note{
There are several ways to interpret the CMO:
\enumerate{
\item Values over/under +/- 50 indicate overbought/oversold conditions.
\item High CMO values indicate strong trends.
\item When the CMO crosses above/below a moving average of the CMO,
it is a buy/sell signal.
}
}
\examples{
\dontrun{
data(TSLA)
# Using default parameters without appending
cmo_result1 <- add_CMO(TSLA)

# Modifying n and without appending
cmo_result2 <- add_CMO(TSLA, n = 20)

# Using default parameters and appending
cmo_result3 <- add_CMO(TSLA, append = TRUE)

# Modifying n and appending
cmo_result4 <- add_CMO(TSLA, n = 20, append = TRUE)
}
}
\references{
The following site(s) were used to code/document this
indicator:\cr \url{https://www.fmlabs.com/reference/CMO.htm}\cr
}
\seealso{
See \code{\link{RSI}}.
}
\author{
DengYishuo
}
\keyword{ts}
