% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_CTI.R
\name{add_CTI}
\alias{add_CTI}
\title{Calculate Ehler's Correlation Trend Indicator}
\usage{
add_CTI(OHLCV, n = 20, slope = 1, append = FALSE)
}
\arguments{
\item{OHLCV}{Price series that is coercible to xts or matrix, assumed to contain Open - High - Low - Close - Volume data.}

\item{n}{Number of periods to use. Defaults to 20.}

\item{slope}{Slope of desired trend. Defaults to 1.}

\item{append}{A logical value. If \code{TRUE}, the calculated Ehler's Correlation Trend Indicator
values will be appended to the \code{OHLCV} input data, ensuring
proper alignment of time - series data. If \code{FALSE}, only the calculated
Ehler's Correlation Trend Indicator values will be returned. Defaults to \code{FALSE}.}
}
\value{
If \code{append = FALSE}, an object of the same class as \code{OHLCV}
or a matrix (if \code{try.xts} fails) with the column:
\describe{
\item{cti}{ The Correlation Trend Indicator. }
}
If \code{append = TRUE}, an object of the same class as \code{OHLCV} with the
calculated Ehler's Correlation Trend Indicator values appended, maintaining the integrity of the time - series
alignment.
}
\description{
Ehler's Correlation Trend Indicator (CTI) measures the Spearman correlation
of the price with the ideal trend line: a straight line with increasing
slope.
The CTI measures the Spearman correlation between the price and the ideal
trend line with slope of \code{slope}, over the past \code{n} days.
See URL in references section for further details.
}
\note{
Positive/negative CTI values signal positive/negative correlation with
the desired trend line slope. A simple strategy could be long when the CTI
is positive and, short when it is negative.
}
\examples{
\dontrun{
data(TSLA)
# Using default parameters without appending
cti_result1 <- add_CTI(TSLA)

# Modifying n and without appending
cti_result2 <- add_CTI(TSLA, n = 30)

# Using default parameters and appending
cti_result3 <- add_CTI(TSLA, append = TRUE)

# Modifying n and appending
cti_result4 <- add_CTI(TSLA, n = 30, append = TRUE)
}
}
\references{
John Ehlers, Correlation Trend Indicator, Stocks & Commodities May - 2020
The following site(s) were used to code/document this indicator:\cr
\url{https://financial - hacker.com/petra - on - programming - a - unique - trend - indicator/}\cr
}
\seealso{
See \code{\link{aroon}}, \code{\link{CCI}}, \code{\link{ADX}},
\code{\link{VHF}}, \code{\link{GMMA}}, \code{\link{TDI}} for other
indicators that measure trend direction/strength.
}
\author{
Ethan Smith, Joshua Ulrich
}
\keyword{ts}
