% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_DonchianChannel.R
\name{add_DonchianChannel}
\alias{add_DonchianChannel}
\alias{ADD_DonchianChannel}
\alias{add_Donchian}
\title{Calculate Donchian Channel}
\usage{
add_DonchianChannel(OHLCV, n = 10, include.lag = FALSE, append = FALSE)
}
\arguments{
\item{OHLCV}{Object that is coercible to xts or matrix and contains Open - High - Low - Close - Volume prices.}

\item{n}{Number of periods for moving average. Defaults to 10.}

\item{include.lag}{Should values be lagged so that today's prices are not
included in the calculation? See Note. Defaults to FALSE.}

\item{append}{A logical value. If \code{TRUE}, the calculated Donchian Channel
values will be appended to the \code{OHLCV} input data, ensuring
proper alignment of time - series data. If \code{FALSE}, only the calculated
Donchian Channel values will be returned. Defaults to \code{FALSE}.}
}
\value{
If \code{append = FALSE}, an object of the same class as \code{OHLCV}
or a matrix (if \code{try.xts} fails) containing the columns:
\describe{
\item{ high }{ The highest high series. }
\item{ mid }{ The average of \code{high} and \code{low}. }
\item{ low }{ The lowest low series. }
}
If \code{append = TRUE}, an object of the same class as \code{OHLCV} with the
calculated Donchian Channel values appended, maintaining the integrity of the time - series
alignment.
}
\description{
Donchian Channels were created by Richard Donchian and were used to generate
buy and sell signals for the Turtle Trading system. Donchian Channels consist
of two (sometimes three) lines: The top line is the highest high of the past
\code{n} periods.  The bottom line is the lowest low of the past \code{n}
periods.  The middle line is the average of the top and bottom lines.
}
\note{
The default of \code{include.lag = FALSE} makes \code{DonchainChannel}
consistent with other \pkg{eTTR} functions, in that it includes the current
period in the calculation.
The default is different than the original calculation, which would calculate
the indicator using periods t - 1 through t - n. Setting \code{include.lag = TRUE}
will return the result of the original calculation. The default of this
argument may change in the future.
}
\examples{
\dontrun{
data(TSLA)
# Using default parameters without appending
dc_result1 <- add_DonchianChannel(TSLA)

# Modifying n and without appending
dc_result2 <- add_DonchianChannel(TSLA, n = 15)

# Using default parameters and appending
dc_result3 <- add_DonchianChannel(TSLA, append = TRUE)

# Modifying n and appending
dc_result4 <- add_DonchianChannel(TSLA, n = 15, append = TRUE)
}
}
\references{
The following site(s) were used to code/document this
indicator:\cr \url{https://www.linnsoft.com/techind/donchian - channels}\cr
}
\seealso{
See \code{\link{BBands}}.
}
\author{
DengYishuo
}
\keyword{ts}
