% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_DVI.R
\name{add_DVI}
\alias{add_DVI}
\title{Calculate DV Intermediate Oscillator}
\usage{
add_DVI(
  OHLCV,
  n = 252,
  wts = c(0.8, 0.2),
  smooth = 3,
  magnitude = c(5, 100, 5),
  stretch = c(10, 100, 2),
  exact.multiplier = 1,
  append = FALSE
)
}
\arguments{
\item{OHLCV}{Price series that is coercible to xts or matrix, assumed to contain Open - High - Low - Close - Volume data.}

\item{n}{Number of periods for the percent rank. Defaults to 252.}

\item{wts}{The weight given to the smoothed returns (magnitude) component and
the up/down days (stretch) component, respectively. Defaults to c(0.8, 0.2).}

\item{smooth}{The number of periods to smooth price. Defaults to 3.}

\item{magnitude}{A set of 3 periods used to smooth magnitude. Defaults to c(5, 100, 5).}

\item{stretch}{A set of 3 periods used to smooth stretch. Defaults to c(10, 100, 2).}

\item{exact.multiplier}{The weight applied to identical values in the window.
See \code{runPercentRank}. Defaults to 1.}

\item{append}{A logical value. If \code{TRUE}, the calculated DV Intermediate Oscillator
values will be appended to the \code{OHLCV} input data, ensuring
proper alignment of time - series data. If \code{FALSE}, only the calculated
DV Intermediate Oscillator values will be returned. Defaults to \code{FALSE}.}
}
\value{
If \code{append = FALSE}, an object of the same class as \code{OHLCV}
or a vector (if \code{try.xts} fails) containing the DVI values.
If \code{append = TRUE}, an object of the same class as \code{OHLCV} with the
calculated DV Intermediate Oscillator values appended, maintaining the integrity of the time - series
alignment.
}
\description{
The DV Intermediate oscillator (DVI) is a very smooth momentum oscillator
that can also be used as a trend indicator.  Created by David Varadi. The DVI
combines smoothed returns over different time windows and the relative number
of up versus down days (stretch) over different time windows.
}
\examples{
\dontrun{
data(TSLA)
# Using default parameters without appending
dvi_result1 <- add_DVI(TSLA)

# Modifying n and without appending
dvi_result2 <- add_DVI(TSLA, n = 300)

# Using default parameters and appending
dvi_result3 <- add_DVI(TSLA, append = TRUE)

# Modifying n and appending
dvi_result4 <- add_DVI(TSLA, n = 300, append = TRUE)
}
}
\references{
The following site(s) were used to code/document this
indicator:\cr
\url{https://cssanalytics.wordpress.com/2009/12/13/what - is - the - dvi/}\cr
\url{https://marketsci.wordpress.com/2010/07/27/css - analytics\%E2\%80\%99 - dvi - indicator - revealed/}\cr
}
\author{
DengYishuo
}
\keyword{ts}
