% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_EMA.R
\name{add_EMA}
\alias{add_EMA}
\title{Calculate Exponential Moving Average (EMA)}
\usage{
add_EMA(OHLCV, n = 10, wilder = FALSE, ratio = NULL, append = FALSE)
}
\arguments{
\item{OHLCV}{Object that is coercible to xts or matrix and contains Open - High - Low - Close - Volume data.
The function will extract the closing price from this object for EMA calculation.}

\item{n}{Number of periods to average over. Must be between 1 and \code{nrow(OHLCV)}.
A larger \code{n} will result in a smoother EMA, but it will be less responsive
to recent price changes.}

\item{wilder}{logical; if \code{TRUE}, use Welles Wilder's EMA formula (\code{1/n}).
When \code{wilder = TRUE}, the smoothing factor is calculated as \code{1/n},
which gives relatively more weight to the most recent data point.}

\item{ratio}{A smoothing/decay ratio (overrides \code{wilder}). If provided,
this ratio is used to calculate the EMA instead of the \code{n}-based approach
(when \code{wilder} is \code{FALSE}) or the Welles Wilder's formula (when
\code{wilder} is \code{TRUE}).}

\item{append}{A logical value. If \code{TRUE}, the calculated EMA values will be appended to the \code{OHLCV} input data,
ensuring proper alignment of time - series data. If \code{FALSE}, only the calculated
EMA values will be returned. Defaults to \code{FALSE}.}
}
\value{
If \code{append} is \code{FALSE}, an object of the same class as \code{OHLCV} (or a vector if \code{try.xts} fails)
containing the EMA values.
If \code{append} is \code{TRUE}, an object of the same class as \code{OHLCV} with the calculated EMA values appended,
maintaining the integrity of the time - series alignment.
}
\description{
Calculate an exponentially - weighted mean, giving more weight to recent observations.
The EMA is a type of moving average that responds more quickly to recent price changes
compared to a simple moving average.
}
\note{
\code{wilder=FALSE} uses \code{2/(n + 1)}; \code{wilder=TRUE} uses \code{1/n}.
}
\examples{
\dontrun{
data(TSLA)
# Using default parameters without appending
ema_result1 <- add_EMA(TSLA)

# Using default parameters and appending
ema_result2 <- add_EMA(TSLA, append = TRUE)

# Changing n and without appending
ema_result3 <- add_EMA(TSLA, n = 15)

# Changing n and appending
ema_result4 <- add_EMA(TSLA, n = 15, append = TRUE)
}
}
\keyword{ts}
