% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_EMV.R
\name{add_EMV}
\alias{add_EMV}
\title{Calculate Arms' Ease of Movement Value}
\usage{
add_EMV(OHLCV, n = 9, maType, vol.divisor = 10000, append = FALSE, ...)
}
\arguments{
\item{OHLCV}{Object that is coercible to xts or matrix and contains Open - High - Low - Close - Volume prices.}

\item{n}{Number of periods for moving average. Defaults to 9.}

\item{maType}{A function or a string naming the function to be called.}

\item{vol.divisor}{An increment to make the results larger and easier to work
with. Defaults to 10000.}

\item{append}{A logical value. If \code{TRUE}, the calculated Arms' Ease of Movement Value
values will be appended to the \code{OHLCV} input data, ensuring
proper alignment of time - series data. If \code{FALSE}, only the calculated
Arms' Ease of Movement Value values will be returned. Defaults to \code{FALSE}.}

\item{...}{Other arguments to be passed to the \code{maType} function.}
}
\value{
If \code{append = FALSE}, an object of the same class as \code{OHLCV}
or a matrix (if \code{try.xts} fails) containing the columns:
\describe{
\item{ emv }{ The ease of movement values. }
\item{ maEMV }{ The smoothed (as specified by \code{ma}) ease of movement values. }
}
If \code{append = TRUE}, an object of the same class as \code{OHLCV} with the
calculated Arms' Ease of Movement Value values appended, maintaining the integrity of the time - series
alignment.
}
\description{
Arms' Ease of Movement Value (EMV) emphasizes days where the security moves
easily and minimizes days where the security does not move easily. Developed
by Richard W. Arms, Jr.
The EMV is calculated by dividing the midpoint \code{[high + low\]/2} move by
the 'Box Ratio' (volume divided by the high minus low).
}
\note{
A buy/sell signal is generated when the EMV crosses above/below zero.
When the EMV hovers around zero, there are small price movements and/or high
volume, and the price is not moving easily.
}
\examples{
\dontrun{
data(TSLA)
# Using default parameters without appending
emv_result1 <- add_EMV(TSLA)

# Modifying n and without appending
emv_result2 <- add_EMV(TSLA, n = 12)

# Using default parameters and appending
emv_result3 <- add_EMV(TSLA, append = TRUE)

# Modifying n and appending
emv_result4 <- add_EMV(TSLA, n = 12, append = TRUE)
}
}
\references{
The following site(s) were used to code/document this
indicator:\cr \url{https://www.fmlabs.com/reference/ArmsEMV.htm}\cr
\url{https://www.metastock.com/Customer/Resources/TAAZ/?p=51}\cr
\url{https://www.linnsoft.com/techind/arms - ease - movement}\cr
}
\seealso{
See \code{\link{EMA}}, \code{\link{SMA}}, etc. for moving average
options; and note Warning section.
}
\author{
DengYishuo
}
\keyword{ts}
