% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_EVWMA.R
\name{add_EVWMA}
\alias{add_EVWMA}
\title{Calculate Elastic Volume - Weighted Moving Average (EVWMA)}
\usage{
add_EVWMA(OHLCV, n = 10, append = FALSE)
}
\arguments{
\item{OHLCV}{Object that is coercible to xts or matrix and contains Open - High - Low - Close - Volume data.
The function will extract the closing price and volume from this object for EVWMA calculation.}

\item{n}{Number of periods to average over. This determines the window size for the moving average calculation.
A larger \code{n} will result in a smoother EVWMA, but it may be less responsive to recent price changes.}

\item{append}{A logical value. If \code{TRUE}, the calculated EVWMA values will be appended to the \code{OHLCV} input data,
ensuring proper alignment of time - series data. If \code{FALSE}, only the calculated
EVWMA values will be returned. Defaults to \code{FALSE}.}
}
\value{
If \code{append} is \code{FALSE}, an object of the same class as \code{OHLCV} (or a vector if \code{try.xts} fails)
containing the EVWMA values.
If \code{append} is \code{TRUE}, an object of the same class as \code{OHLCV} with the calculated EVWMA values appended,
maintaining the integrity of the time - series alignment.
}
\description{
Calculate a volume - weighted moving average that adjusts to trading volume.
The EVWMA takes into account the trading volume when calculating the moving average,
giving more importance to periods with higher volume.
}
\examples{
\dontrun{
data(TSLA)
# Using default parameters without appending
evwma_result1 <- add_EVWMA(TSLA)

# Using default parameters and appending
evwma_result2 <- add_EVWMA(TSLA, append = TRUE)

# Changing n and without appending
evwma_result3 <- add_EVWMA(TSLA, n = 15)

# Changing n and appending
evwma_result4 <- add_EVWMA(TSLA, n = 15, append = TRUE)
}
}
\keyword{ts}
