% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_GMMA.R
\name{add_GMMA}
\alias{add_GMMA}
\alias{ADD_GMMA}
\title{Calculate Guppy Multiple Moving Averages}
\usage{
add_GMMA(
  OHLCV,
  short = c(3, 5, 8, 10, 12, 15),
  long = c(30, 35, 40, 45, 50, 60),
  maType,
  append = FALSE
)
}
\arguments{
\item{OHLCV}{Price, volume, etc. series that is coercible to xts or matrix, assumed to contain Open - High - Low - Close - Volume data.}

\item{short}{Vector of short - term periods. Defaults to c(3, 5, 8, 10, 12, 15).}

\item{long}{Vector of long - term periods. Defaults to c(30, 35, 40, 45, 50, 60).}

\item{maType}{Either:
\enumerate{
\item A function or a string naming the function to be called.
\item A \emph{list} with the first component like (1) above, and
additional parameters specified as \emph{named} components.
See Examples.
}}

\item{append}{A logical value. If \code{TRUE}, the calculated Guppy Multiple Moving Average
values will be appended to the \code{OHLCV} input data, ensuring
proper alignment of time - series data. If \code{FALSE}, only the calculated
Guppy Multiple Moving Average values will be returned. Defaults to \code{FALSE}.}
}
\value{
If \code{append = FALSE}, an object of the same class as \code{OHLCV}
or a vector (if \code{try.xts} fails) containing the Guppy Multiple Moving Average.
If \code{append = TRUE}, an object of the same class as \code{OHLCV} with the
calculated Guppy Multiple Moving Average values appended, maintaining the integrity of the time - series
alignment.
}
\description{
Calculate the Guppy Multiple Moving Average of a series.
The Guppy Multiple Moving Average signals a changing trend when the
\code{short} and \code{long} groups of moving averages intersect.  An up/down
trend exists when the short/long - term moving averages are greater than the
long/short - term averages.
}
\examples{
\dontrun{
data(TSLA)
# Using default parameters without appending
gmma_result1 <- add_GMMA(TSLA)

# Using default parameters and appending
gmma_result2 <- add_GMMA(TSLA, append = TRUE)
}
}
\references{
The following site(s) were used to code/document this
indicator:\cr
\url{https://www.investopedia.com/terms/g/guppy - multiple - moving - average.asp}\cr
}
\seealso{
See \code{\link{aroon}}, \code{\link{CCI}}, \code{\link{ADX}},
\code{\link{VHF}}, \code{\link{TDI}} for other indicators that measure trend
direction/strength.
}
\author{
DengYishuo
}
\keyword{ts}
