% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_keltnerChannels.R
\name{add_keltnerChannels}
\alias{add_keltnerChannels}
\alias{add_KeltnerChannels}
\title{Calculate Keltner Channels}
\usage{
add_keltnerChannels(OHLCV, n = 20, maType, atr = 2, append = FALSE, ...)
}
\arguments{
\item{OHLCV}{Object that is coercible to xts or matrix and contains Open - High - Low - Close - Volume prices.}

\item{n}{Number of periods for moving average. Defaults to 20.}

\item{maType}{A function or a string naming the function to be called.}

\item{atr}{The number of average true range distances to apply. Defaults to 2.}

\item{append}{A logical value. If \code{TRUE}, the calculated Keltner Channels
values will be appended to the \code{OHLCV} input data, ensuring
proper alignment of time - series data. If \code{FALSE}, only the calculated
Keltner Channels values will be returned. Defaults to \code{FALSE}.}

\item{...}{Other arguments to be passed to the maType function.}
}
\value{
If \code{append = FALSE}, an object of the same class as \code{OHLCV}
or a matrix (if \code{try.xts} fails) containing the columns:
\describe{
\item{dn}{ The lower Keltner Channel. }
\item{mavg}{ The middle moving average. }
\item{up}{ The upper Keltner Channel. }
}
If \code{append = TRUE}, an object of the same class as \code{OHLCV} with the
calculated Keltner Channels values appended, maintaining the integrity of the time - series
alignment.
}
\description{
Keltner Channels are volatility - based envelopes set above and below a moving
average. This indicator is similar to Bollinger Bands, but Keltner Channels
use the Average True Range (ATR) to set channel distance.
Keltner Channels are a trend following indicator, and can also be used to
identify overbought and oversold levels when there is no trend.
Chester Keltner is credited with the original version of Keltner Channels in
his 1960 book. Linda Bradford Raschke introduced the newer version of
Keltner Channels in the 1980s.
}
\section{Details }{
 Keltner Channels consist of three lines:
The middle band is generally a 20 - period EMA of the typical price
\code{[high + low + close]/3}. The upper and lower bands are multiples of average
true range (usually 2) above and below the MA.
The middle band is usually calculated using the typical price, but if a
univariate series (e.g. Close, Weighted Close, Median Price, etc.) is
provided, it will be used instead.
}

\examples{
\dontrun{
data(TSLA)
# Using default parameters without appending
kc_result1 <- add_keltnerChannels(TSLA)

# Modifying n and without appending
kc_result2 <- add_keltnerChannels(TSLA, n = 30)

# Using default parameters and appending
kc_result3 <- add_keltnerChannels(TSLA, append = TRUE)

# Modifying n and appending
kc_result4 <- add_keltnerChannels(TSLA, n = 30, append = TRUE)
}
}
\references{
The following site(s) were used to code/document this
indicator:\cr
\url{https://school.stockcharts.com/doku.php?id=technical_indicators:keltner_channels}\cr
\url{https://www.linnsoft.com/techind/keltner - channels - keltu - keltd}\cr
\url{https://www.investopedia.com/terms/k/keltnerchannel.asp}\cr
}
\seealso{
See \code{\link{EMA}}, \code{\link{SMA}}, etc. for moving average
options; and note Warning section.
}
\author{
Nick Procyk, Joshua Ulrich
}
\keyword{ts}
