% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_KST.R
\name{add_KST}
\alias{add_KST}
\title{Calculate Know Sure Thing}
\usage{
add_KST(
  OHLCV,
  n = c(10, 10, 10, 15),
  nROC = c(10, 15, 20, 30),
  nSig = 9,
  maType,
  wts = 1:NROW(n),
  append = FALSE,
  ...
)
}
\arguments{
\item{OHLCV}{Price series that is coercible to xts or matrix, assumed to contain Open - High - Low - Close - Volume data.}

\item{n}{A vector of the number of periods to use in the MA calculations. Defaults to c(10, 10, 10, 15).}

\item{nROC}{A vector of the number of periods to use in the ROC calculations. Defaults to c(10, 15, 20, 30).}

\item{nSig}{The number of periods to use for the KST signal line. Defaults to 9.}

\item{maType}{Either:
\enumerate{
\item A function or a string naming the function to be called.
\item A \emph{list} with the first component like (1) above, and
additional parameters specified as \emph{named} components.
See Examples.
}}

\item{wts}{A vector the same length as \code{n}, of the weight for each
period (need not sum to one). Defaults to 1:NROW(n).}

\item{append}{A logical value. If \code{TRUE}, the calculated Know Sure Thing
values will be appended to the \code{OHLCV} input data, ensuring
proper alignment of time - series data. If \code{FALSE}, only the calculated
Know Sure Thing values will be returned. Defaults to \code{FALSE}.}

\item{...}{Other arguments to be passed to the \code{maType} function in
case (1) above.}
}
\value{
If \code{append = FALSE}, an object of the same class as \code{OHLCV}
or a vector (if \code{try.xts} fails) containing the Know Sure Thing values.
If \code{append = TRUE}, an object of the same class as \code{OHLCV} with the
calculated Know Sure Thing values appended, maintaining the integrity of the time - series
alignment.
}
\description{
The Know Sure Thing (KST) is a smooth, summed, rate of change indicator.
Developed by Martin Pring.
For each day (week, month, etc.), the KST calculates the ROC over several
periods. Those ROCs are smoothed using the given moving averages, then
multiplied by their respective weighting values. The resulting values are
summed for each day (month, week, etc.).
}
\note{
The KST indicates bullish/bearish momentum as it crosses above/below
its moving average. Because the KST tends to lead price action, look for
trend confirmation in the price.
The default arguments are for the daily KST. There is also the Long - Term
KST, with arguments: \code{n=c(9, 12, 18, 24)} - where the periods are
months, not days - and the moving average periods are 6, 6, 6, and 9 months,
respectively.
}
\examples{
\dontrun{
data(TSLA)
# Using default parameters without appending
kst_result1 <- add_KST(TSLA)

# Modifying n and without appending
kst_result2 <- add_KST(TSLA, n = c(12, 12, 12, 18))

# Using default parameters and appending
kst_result3 <- add_KST(TSLA, append = TRUE)

# Modifying n and appending
kst_result4 <- add_KST(TSLA, n = c(12, 12, 12, 18), append = TRUE)
}
}
\references{
The following site(s) were used to code/document this
indicator:\cr
\url{https://web.archive.org/web/20110715112957/http://www.pring.com/movieweb/daily_kst.htm}\cr
\url{https://web.archive.org/web/20100101162707/http://www.pring.com/movieweb/KST_MCM.htm}\cr
}
\seealso{
See \code{\link{EMA}}, \code{\link{SMA}}, etc. for moving average
options; and note Warning section.  See \code{\link{ROC}} for the
rate - of - change function.  See \code{\link{MACD}} for a generic oscillator.
}
\author{
DengYishuo
}
\keyword{ts}
