% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_MFI.R
\name{add_MFI}
\alias{add_MFI}
\alias{ADD_MFI}
\alias{add_mfi}
\title{Calculate Money Flow Index}
\usage{
add_MFI(OHLCV, n = 14, append = FALSE)
}
\arguments{
\item{OHLCV}{Object that is coercible to xts or matrix and contains Open - High - Low - Close - Volume prices.}

\item{n}{Number of periods to use. Defaults to 14.}

\item{append}{A logical value. If \code{TRUE}, the calculated Money Flow Index
values will be appended to the \code{OHLCV} input data, ensuring
proper alignment of time - series data. If \code{FALSE}, only the calculated
Money Flow Index values will be returned. Defaults to \code{FALSE}.}
}
\value{
If \code{append = FALSE}, an object of the same class as \code{OHLCV}
or a vector (if \code{try.xts} fails) containing the MFI values.
If \code{append = TRUE}, an object of the same class as \code{OHLCV} with the
calculated Money Flow Index values appended, maintaining the integrity of the time - series
alignment.
}
\description{
The MFI is a ratio of positive and negative money flow over time.
Money Flow (MF) is the product of price and volume. Positive/negative MF
occur when today's price is higher/lower than yesterday's price. The MFI is
calculated by dividing positive MF by negative MF for the past \code{n}
periods. It is then scaled between 0 and 100.
MFI is usually calculated using the typical price, but if a univariate series
(e.g. Close, Weighted Close, Median Price, etc.) is provided, it will be used
instead.
}
\note{
Divergence between MFI and price can be indicative of a reversal. In
addition, values above/below 80/20 indicate market tops/bottoms.
}
\examples{
\dontrun{
data(TSLA)
# Using default parameters without appending
mfi_result1 <- add_MFI(TSLA)

# Modifying n and without appending
mfi_result2 <- add_MFI(TSLA, n = 20)

# Using default parameters and appending
mfi_result3 <- add_MFI(TSLA, append = TRUE)

# Modifying n and appending
mfi_result4 <- add_MFI(TSLA, n = 20, append = TRUE)
}
}
\references{
The following site(s) were used to code/document this
indicator:\cr
\url{https://www.fmlabs.com/reference/default.htm?url=MoneyFlowIndex.htm}\cr
\url{https://www.linnsoft.com/techind/money - flow - index - mfi}\cr
\url{https://school.stockcharts.com/doku.php?id=technical_indicators:money_flow_index_mfi}\cr
}
\seealso{
See \code{\link{OBV}} and \code{\link{CMF}}.
}
\author{
DengYishuo
}
\keyword{ts}
