% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_OBV.R
\name{add_OBV}
\alias{add_OBV}
\title{Calculate On Balance Volume (OBV)}
\usage{
add_OBV(OHLCV, append = FALSE)
}
\arguments{
\item{OHLCV}{Object that is coercible to xts or matrix and contains Open - High - Low - Close - Volume data.
The function will extract the closing price and volume from this object for OBV calculation.}

\item{append}{A logical value. If \code{TRUE}, the calculated OBV values will be appended to the \code{OHLCV} input data,
ensuring proper alignment of time - series data. If \code{FALSE}, only the calculated
OBV values will be returned. Defaults to \code{FALSE}.}
}
\value{
If \code{append} is \code{FALSE}, an object of the same class as \code{OHLCV} (or a vector if \code{try.xts} fails)
containing the OBV values.
If \code{append} is \code{TRUE}, an object of the same class as \code{OHLCV} with the calculated OBV values appended,
maintaining the integrity of the time - series alignment.
}
\description{
On Balance Volume (OBV) is a momentum - based technical indicator that measures
the cumulative flow of volume to determine the strength of a price trend.
It uses volume flow to predict changes in stock prices by adding volume on
up days and subtracting volume on down days.
}
\note{
OBV is typically used to confirm price trends or identify potential
divergences between volume flow and price movements.
}
\examples{
\dontrun{
data(TSLA)
# Using default parameters without appending
obv_result1 <- add_OBV(TSLA)

# Using default parameters and appending
obv_result2 <- add_OBV(TSLA, append = TRUE)
}
}
\references{
\itemize{
\item https://www.fmlabs.com/reference/OBV.htm
\item https://www.metastock.com/Customer/Resources/TAAZ/?p=82
\item https://www.linnsoft.com/techind/balance - open - interest
\item https://school.stockcharts.com/doku.php?id=technical_indicators:on_balance_volume_obv
}
}
\seealso{
See \code{\link{chaikinAD}} for a similar volume - based indicator.
}
\author{
DengYishuo
}
\keyword{indicator}
\keyword{momentum}
\keyword{ts}
\keyword{volume}
