% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_PBands.R
\name{add_PBands}
\alias{add_PBands}
\alias{add_pbands}
\alias{add_priceBands}
\title{Construct (optionally further smoothed and centered ) volatility bands around prices}
\usage{
add_PBands(
  OHLCV,
  n = 20,
  maType = "SMA",
  sd = 2,
  fastn = 2,
  centered = FALSE,
  lavg = FALSE,
  append = FALSE,
  ...
)
}
\arguments{
\item{OHLCV}{Object that is coercible to xts or matrix, assumed to contain Open - High - Low - Close - Volume data.}

\item{n}{Number of periods to average over. Defaults to 20.}

\item{maType}{A function or a string naming the function to be called. Defaults to "SMA".}

\item{sd}{The number of standard deviations to use. Defaults to 2.}

\item{fastn}{Number of periods to use for smoothing higher - frequency 'noise'. Defaults to 2.}

\item{centered}{Whether to center the bands around a series adjusted for high
frequency noise, default \code{FALSE}.}

\item{lavg}{Whether to use a longer \code{(n*2)} smoothing period for
centering, default \code{FALSE}.}

\item{append}{A logical value. If \code{TRUE}, the calculated price volatility bands
values will be appended to the \code{OHLCV} input data, ensuring
proper alignment of time - series data. If \code{FALSE}, only the calculated
price volatility bands values will be returned. Defaults to \code{FALSE}.}

\item{...}{any other pass - thru parameters, usually for function named by
\code{maType}.}
}
\value{
If \code{append = FALSE}, an object of the same class as \code{OHLCV}
or a matrix (if \code{try.xts} fails) containing the columns:
\describe{
\item{ dn }{ The lower price volatility Band. }
\item{ center }{ The smoothed centerline (see details). }
\item{ up }{ The upper price volatility Band. }
}
If \code{append = TRUE}, an object of the same class as \code{OHLCV} with the
calculated price volatility bands values appended, maintaining the integrity of the time - series
alignment.
}
\description{
John Bollinger's famous adaptive volatility bands most often use the typical
price of an HLC series, or may be calculated on a univariate price series
(see \code{\link{BBands}}).
This function applies a second moving average denoted by \code{fastn} to
filter out higher - frequency noise, making the bands somewhat more stable to
temporary fluctuations and spikes.
If \code{centered} is \code{TRUE}, the function also further smoothes and
centers the bands around a centerline adjusted to remove this higher
frequency noise.  If \code{lavg} is also \code{TRUE}, the smoothing applied
for the middle band (but not the volatility bands) is doubled to further
smooth the price - response function.
If you have multiple different price series in \code{prices}, and want to use
this function, call this functions using \code{lapply(prices,PBands,...)}.
}
\examples{
\dontrun{
data(TSLA)
# Using default parameters without appending
pbands_result1 <- add_PBands(TSLA)

# Modifying n and without appending
pbands_result2 <- add_PBands(TSLA, n = 25)

# Using default parameters and appending
pbands_result3 <- add_PBands(TSLA, append = TRUE)

# Modifying n and appending
pbands_result4 <- add_PBands(TSLA, n = 25, append = TRUE)
}
}
\seealso{
\code{\link{BBands}}
}
\author{
Brian G. Peterson
}
\keyword{ts}
