% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_PO.R
\name{add_PO}
\alias{add_PO}
\title{Calculate Price Oscillator (PO)}
\usage{
add_PO(
  OHLCV,
  n_short = 12,
  n_long = 26,
  type = c("difference", "percent"),
  ma_type = c("SMA", "EMA"),
  append = FALSE
)
}
\arguments{
\item{OHLCV}{An object that is coercible to xts or matrix, assumed to contain Open - High - Low - Close - Volume data.}

\item{n_short}{The period for the short - term moving average, default is 12.}

\item{n_long}{The period for the long - term moving average, default is 26.}

\item{type}{The type of return value: "difference" for the absolute difference,
"percent" for the percentage difference. Default is "difference".}

\item{ma_type}{The type of moving average, supports "SMA" (Simple Moving Average)
and "EMA" (Exponential Moving Average). Default is "SMA".}

\item{append}{A logical value. If \code{TRUE}, the calculated Price Oscillator
values will be appended to the \code{OHLCV} input data, ensuring
proper alignment of time - series data. If \code{FALSE}, only the calculated
Price Oscillator values will be returned. Defaults to \code{FALSE}.}
}
\value{
If \code{append = FALSE}, a PO indicator series of the same type as the input price series (extracted from OHLCV).
If \code{append = TRUE}, an object of the same class as \code{OHLCV} with the
calculated Price Oscillator values appended, maintaining the integrity of the time - series
alignment.
}
\description{
Calculates the Price Oscillator (PO), a momentum indicator that measures the difference
between two moving averages of different periods. It helps identify trends and potential
reversal points in financial markets.
}
\details{
The Price Oscillator (PO) measures momentum by calculating the difference between
a short - term and a long - term moving average of prices. A positive PO value indicates
upward momentum, while a negative value suggests downward momentum.
}
\examples{
\dontrun{
data(TSLA)
# Using default parameters without appending
po_result1 <- add_PO(TSLA)

# Modifying n_short and without appending
po_result2 <- add_PO(TSLA, n_short = 14)

# Using default parameters and appending
po_result3 <- add_PO(TSLA, append = TRUE)

# Modifying n_short and appending
po_result4 <- add_PO(TSLA, n_short = 14, append = TRUE)
}
}
\references{
Murphy, J. (1999). Technical Analysis of the Financial Markets. New York Institute of Finance.
}
\seealso{
\code{\link{SMA}}, \code{\link{EMA}} for moving average calculations.
}
