% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_RSI.R
\name{add_RSI}
\alias{add_RSI}
\title{Calculate Relative Strength Index}
\usage{
add_RSI(OHLCV, n = 14, maType, append = FALSE, ...)
}
\arguments{
\item{OHLCV}{Object that is coercible to xts or matrix, assumed to contain Open - High - Low - Close - Volume data.}

\item{n}{Number of periods for moving averages. Defaults to 14.}

\item{maType}{Either:
\enumerate{
\item A function or a string naming the function to be called.
\item A \emph{list} with the first component like (1) above, and
additional parameters specified as \emph{named} components.
See Examples.
}}

\item{append}{A logical value. If \code{TRUE}, the calculated Relative Strength Index
values will be appended to the \code{OHLCV} input data, ensuring
proper alignment of time - series data. If \code{FALSE}, only the calculated
Relative Strength Index values will be returned. Defaults to \code{FALSE}.}

\item{...}{Other arguments to be passed to the \code{maType} function in
case (1) above.}
}
\value{
If \code{append = FALSE}, an object of the same class as \code{OHLCV}
or a vector (if \code{try.xts} fails) containing the RSI values.
If \code{append = TRUE}, an object of the same class as \code{OHLCV} with the
calculated Relative Strength Index values appended, maintaining the integrity of the time - series
alignment.
}
\description{
The Relative Strength Index (RSI) calculates a ratio of the recent upward
price movements to the absolute price movement. Developed by J. Welles
Wilder.
The RSI calculation is \code{RSI = 100 - 100 / ( 1 + RS )}, where \code{RS}
is the smoothed ratio of 'average' gains over 'average' losses. The
'averages' aren't true averages, since they're divided by the value of
\code{n} and not the number of periods in which there are  gains/losses.
}
\note{
The RSI is usually interpreted as an overbought/oversold (over 70 /
below 30) indicator. Divergence with price may also be useful. For example,
if price is making new highs/lows, but RSI is not, it could indicate a
reversal.
You can calculate a stochastic RSI by using the function \code{\link{stoch}}
on RSI values.
}
\examples{
\dontrun{
data(TSLA)
# Using default parameters without appending
rsi_result1 <- add_RSI(TSLA)

# Modifying n and without appending
rsi_result2 <- add_RSI(TSLA, n = 20)

# Using default parameters and appending
rsi_result3 <- add_RSI(TSLA, append = TRUE)

# Modifying n and appending
rsi_result4 <- add_RSI(TSLA, n = 20, append = TRUE)
}
}
\references{
The following site(s) were used to code/document this
indicator:
\cr Relative Strength Index:\cr
\url{https://www.fmlabs.com/reference/RSI.htm}\cr
\url{https://www.metastock.com/Customer/Resources/TAAZ/?p=100}\cr
\url{https://www.linnsoft.com/techind/relative - strength - index - rsi}\cr
\url{https://school.stockcharts.com/doku.php?id=technical_indicators:relative_strength_index_rsi}\cr
\cr Stochastic RSI:\cr
\url{https://www.fmlabs.com/reference/StochRSI.htm}\cr
\url{https://school.stockcharts.com/doku.php?id=technical_indicators:stochrsi}\cr
}
\seealso{
See \code{\link{EMA}}, \code{\link{SMA}}, etc. for moving average
options; and note Warning section. See \code{\link{CMO}} for a variation on
RSI.
}
\author{
DengYishuo
}
\keyword{ts}
