% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_SAR.R
\name{add_SAR}
\alias{add_SAR}
\title{Calculate Parabolic Stop - and - Reverse}
\usage{
add_SAR(OHLCV, accel = c(0.02, 0.2), append = FALSE)
}
\arguments{
\item{OHLCV}{Object that is coercible to xts or matrix, assumed to contain Open - High - Low - Close - Volume data.}

\item{accel}{A vector where \code{accel[1]} is the initial acceleration factor
and \code{accel[2]} is the maximum acceleration factor. Defaults to c(.02, .2).}

\item{append}{A logical value. If \code{TRUE}, the calculated Parabolic Stop - and - Reverse
values will be appended to the \code{OHLCV} input data, ensuring
proper alignment of time - series data. If \code{FALSE}, only the calculated
Parabolic Stop - and - Reverse values will be returned. Defaults to \code{FALSE}.}
}
\value{
If \code{append = FALSE}, an object of the same class as \code{OHLCV}
or a vector (if \code{try.xts} fails) containing the Parabolic Stop and Reverse values.
If \code{append = TRUE}, an object of the same class as \code{OHLCV} with the
calculated Parabolic Stop and Reverse values appended, maintaining the integrity of the time - series
alignment.
}
\description{
The Parabolic Stop - and - Reverse calculates a trailing stop. Developed by J.
Welles Wilder.
The calculation for the SAR is quite complex. See the URLs in the references
section for calculation notes.
The SAR assumes that you are always in the market, and calculates the Stop
And Reverse point when you would close a long position and open a short
position or vice versa.
}
\examples{
\dontrun{
data(TSLA)
# Using default parameters without appending
sar_result1 <- add_SAR(TSLA)

# Modifying accel and without appending
sar_result2 <- add_SAR(TSLA, accel = c(0.03, 0.25))

# Using default parameters and appending
sar_result3 <- add_SAR(TSLA, append = TRUE)

# Modifying accel and appending
sar_result4 <- add_SAR(TSLA, accel = c(0.03, 0.25), append = TRUE)
}
}
\references{
The following site(s) were used to code/document this
indicator:\cr
\url{https://www.linnsoft.com/techind/parabolic - sar - sar}\cr
\url{https://www.fmlabs.com/reference/SAR.htm}\cr
\url{https://school.stockcharts.com/doku.php?id=technical_indicators:parabolic_sar}\cr
\url{https://www.metastock.com/Customer/Resources/TAAZ/?p=87}
}
\seealso{
See \code{\link{ATR}} and \code{\link{ADX}}, which were also
developed by Welles Wilder.
}
\author{
DengYishuo
}
\keyword{ts}
