% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_SMA.R
\name{add_SMA}
\alias{add_SMA}
\title{Calculate Simple Moving Average (SMA)}
\usage{
add_SMA(OHLCV, n = 10, append = FALSE)
}
\arguments{
\item{OHLCV}{Object that is coercible to xts or matrix and contains Open - High - Low - Close - Volume data.
The function will extract the closing price from this object for SMA calculation.}

\item{n}{Number of periods to average over. Must be between 1 and \code{nrow(OHLCV)}.
A larger \code{n} results in a smoother SMA, but it may be less responsive to recent price changes.}

\item{append}{A logical value. If \code{TRUE}, the calculated SMA values will be appended to the \code{OHLCV} input data,
ensuring proper alignment of time - series data. If \code{FALSE}, only the calculated
SMA values will be returned. Defaults to \code{FALSE}.}
}
\value{
If \code{append} is \code{FALSE}, an object of the same class as \code{OHLCV} (or a vector if \code{try.xts} fails)
containing the SMA values.
If \code{append} is \code{TRUE}, an object of the same class as \code{OHLCV} with the calculated SMA values appended,
maintaining the integrity of the time - series alignment.
}
\description{
Calculate the arithmetic mean of the series over the past \code{n} observations.
The Simple Moving Average (SMA) is a basic technical analysis tool that provides
a smoothed view of the price trend by averaging the closing prices over a
specified number of periods.
}
\examples{
\dontrun{
data(TSLA)
# Using default parameters without appending
sma_result1 <- add_SMA(TSLA)

# Using default parameters and appending
sma_result2 <- add_SMA(TSLA, append = TRUE)

# Changing n and without appending
sma_result3 <- add_SMA(TSLA, n = 15)

# Changing n and appending
sma_result4 <- add_SMA(TSLA, n = 15, append = TRUE)
}
}
\keyword{ts}
