% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_SMI.R
\name{add_SMI}
\alias{add_SMI}
\title{Calculate Stochastic Momentum Index}
\usage{
add_SMI(
  OHLCV,
  n = 13,
  nFast = 2,
  nSlow = 25,
  nSig = 9,
  maType,
  bounded = TRUE,
  append = FALSE,
  ...
)
}
\arguments{
\item{OHLCV}{Object coercible to xts or matrix, assumed to contain Open - High - Low - Close - Volume data.}

\item{n}{Number of periods for range calculation (default: 13).}

\item{nFast}{Fast moving average periods (default: 2).}

\item{nSlow}{Slow moving average periods (default: 25).}

\item{nSig}{Signal line periods (default: 9).}

\item{maType}{Moving average type: function, string, or list of configurations
(default: "EMA").}

\item{bounded}{Logical, use current period in range calculation (default: TRUE).}

\item{append}{A logical value. If \code{TRUE}, the calculated Stochastic Momentum Index
values and its signal line will be appended to the \code{OHLCV} input data, ensuring
proper alignment of time - series data. If \code{FALSE}, only the calculated
Stochastic Momentum Index values and its signal line will be returned. Defaults to \code{FALSE}.}

\item{...}{Additional arguments passed to moving average functions.}
}
\value{
If \code{append = FALSE}, an xts or matrix object with columns: SMI, signal.
If \code{append = TRUE}, an object of the same class as \code{OHLCV} with the
calculated Stochastic Momentum Index values and its signal line appended, maintaining the integrity of the time - series
alignment.
}
\description{
Calculates the SMI value and its signal line.
}
\examples{
\dontrun{
data(TSLA)
# Using default parameters without appending
smi_result1 <- add_SMI(TSLA)

# Modifying n and without appending
smi_result2 <- add_SMI(TSLA, n = 15)

# Using default parameters and appending
smi_result3 <- add_SMI(TSLA, append = TRUE)

# Modifying n and appending
smi_result4 <- add_SMI(TSLA, n = 15, append = TRUE)
}
}
\author{
DengYishuo
}
\keyword{indicator}
\keyword{momentum}
\keyword{ts}
