% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_SNR.R
\name{add_SNR}
\alias{add_SNR}
\title{Calculate Signal to Noise Ratio}
\usage{
add_SNR(OHLCV, n, append = FALSE, ...)
}
\arguments{
\item{OHLCV}{Object that is coercible to xts or matrix, assumed to contain Open - High - Low - Close - Volume data.}

\item{n}{Number of periods for moving average.}

\item{append}{A logical value. If \code{TRUE}, the calculated Signal to Noise Ratio
values will be appended to the \code{OHLCV} input data, ensuring
proper alignment of time - series data. If \code{FALSE}, only the calculated
Signal to Noise Ratio values will be returned. Defaults to \code{FALSE}.}

\item{...}{Other arguments to be passed to \code{\link{ATR}}.}
}
\value{
If \code{append = FALSE}, an object of the same class as \code{OHLCV}
or a matrix (if \code{try.xts} fails) containing the signal to noise ratio.
If \code{append = TRUE}, an object of the same class as \code{OHLCV} with the
calculated Signal to Noise Ratio values appended, maintaining the integrity of the time - series
alignment.
}
\description{
The n - day SNR for a given market is calculated by taking the absolute
price change over an n - day period and dividing it by the average
n - day volatility.

\deqn{SNR_n = \frac{|C_t - C_{t - n}|}{ATR_n}
}{SNR = abs(Cl - lag(Cl,n)) / ATR(HLC, n)$atr}

Using average true range as the volatility measure captures more of the
intraday and overnight volatility in a way that a measurement of
Close - to - Close price change does not.

The interpretation is then relatively intuitive: an SNR value of five
indicates that the market has moved five times the volatility (average true
range) over the given look - back period.
}
\references{
Skeggs, James and Hill, Alex (2015). Back in Black Part 2: The
Opportunity Set for Trend Following.
}
\author{
Peter Carl
}
