% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_stoch.R
\name{add_stoch}
\alias{add_stoch}
\title{Calculate Stochastic Oscillator}
\usage{
add_stoch(
  OHLCV,
  nFastK = 14,
  nFastD = 3,
  nSlowD = 3,
  maType,
  bounded = TRUE,
  smooth = 1,
  append = FALSE,
  ...
)
}
\arguments{
\item{OHLCV}{Object coercible to xts or matrix, assumed to contain Open - High - Low - Close - Volume data.}

\item{nFastK}{Number of periods for fast \%K calculation (default: 14).}

\item{nFastD}{Number of periods for fast \%D calculation (default: 3).}

\item{nSlowD}{Number of periods for slow \%D calculation (default: 3).}

\item{maType}{Moving average type: function, string, or list of configurations
(default: "SMA").}

\item{bounded}{Logical, use current period in range calculation (default: TRUE).}

\item{smooth}{Internal smoothing periods for FastK (default: 1).}

\item{append}{A logical value. If \code{TRUE}, the calculated Stochastic Oscillator
values (fast \%K, fast \%D, slow \%D) will be appended to the \code{OHLCV} input data, ensuring
proper alignment of time - series data. If \code{FALSE}, only the calculated
Stochastic Oscillator values will be returned. Defaults to \code{FALSE}.}

\item{...}{Additional arguments passed to moving average functions.}
}
\value{
If \code{append = FALSE}, an xts or matrix object with columns: fastK, fastD, slowD.
If \code{append = TRUE}, an object of the same class as \code{OHLCV} with the
calculated Stochastic Oscillator values appended, maintaining the integrity of the time - series
alignment.
}
\description{
Calculates the fast \%K, fast \%D, and slow \%D values of
the stochastic oscillator.
}
\examples{
\dontrun{
data(TSLA)
# Using default parameters without appending
stoch_result1 <- add_stoch(TSLA)

# Modifying nFastK and without appending
stoch_result2 <- add_stoch(TSLA, nFastK = 10)

# Using default parameters and appending
stoch_result3 <- add_stoch(TSLA, append = TRUE)

# Modifying nFastK and appending
stoch_result4 <- add_stoch(TSLA, nFastK = 10, append = TRUE)
}
}
\author{
DengYishuo
}
\keyword{indicator}
\keyword{momentum}
\keyword{ts}
