% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_TDI.R
\name{add_TDI}
\alias{add_TDI}
\title{Calculate Trend Detection Index}
\usage{
add_TDI(OHLCV, n = 20, multiple = 2, allow_middle_na = FALSE, append = FALSE)
}
\arguments{
\item{OHLCV}{Object that is coercible to xts or matrix, assumed to contain Open - High - Low - Close - Volume data.}

\item{n}{Number of periods to use. Default is 20.}

\item{multiple}{Multiple used to calculate (2). Default is 2.}

\item{allow_middle_na}{Logical. If \code{TRUE}, allows NA values in the middle. Default is FALSE.}

\item{append}{A logical value. If \code{TRUE}, the calculated Trend Detection Index
and Direction Indicator values will be appended to the \code{OHLCV} input data, ensuring
proper alignment of time - series data. If \code{FALSE}, only the calculated
Trend Detection Index and Direction Indicator values will be returned. Defaults to \code{FALSE}.}
}
\value{
If \code{append = FALSE}, an object of the same class as \code{OHLCV}
or a matrix (if \code{try.xts} fails) containing the columns:
\describe{
\item{ tdi }{ The Trend Detection Index. }
\item{ di }{ The Direction Indicator. }
}
If \code{append = TRUE}, an object of the same class as \code{OHLCV} with the
calculated Trend Detection Index and Direction Indicator values appended, maintaining the integrity of the time - series
alignment.
}
\description{
The Trend Detection Index (TDI) attempts to identify starting and ending
trends. Developed by M. H. Pee.
The TDI is the (1) absolute value of the \code{n}-day sum of the \code{n}-day
momentum, minus the quantity of (2) \code{multiple}*\code{n}-day sum of the
absolute value of the \code{n}-day momentum, minus (3) \code{n}-day sum of
the absolute value of the \code{n}-day momentum.
I.e. \eqn{TDI = (1) - [ (2) - (3) ]}
The direction indicator is the sum of the \code{n}-day momentum over the last
\code{n} days.
See URL in references section for further details.
}
\note{
Positive/negative TDI values signal a trend/consolidation. A positive/
negative direction indicator signals a up/down trend. I.e. buy if the TDI
and the direction indicator are positive, and sell if the TDI is positive
while the direction indicator is negative.
}
\examples{
\dontrun{
data(TSLA)
# Using default parameters without appending
tdi_result1 <- add_TDI(TSLA)

# Modifying n and without appending
tdi_result2 <- add_TDI(TSLA, n = 25)

# Using default parameters and appending
tdi_result3 <- add_TDI(TSLA, append = TRUE)

# Modifying n and appending
tdi_result4 <- add_TDI(TSLA, n = 25, append = TRUE)
}
}
\references{
The following site(s) were used to code/document this
indicator:\cr
\url{https://www.linnsoft.com/techind/trend - detection - index - tdi}\cr
}
\seealso{
See \code{\link{aroon}}, \code{\link{CCI}}, \code{\link{ADX}},
\code{\link{VHF}}, \code{\link{GMMA}} for other indicators that measure trend
direction/strength.
}
\author{
DengYishuo
}
\keyword{ts}
