% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_TRIX.R
\name{add_TRIX}
\alias{add_TRIX}
\title{Calculate Triple Smoothed Exponential Oscillator}
\usage{
add_TRIX(OHLCV, n = 20, nSig = 9, maType, percent = TRUE, append = FALSE, ...)
}
\arguments{
\item{OHLCV}{Object that is coercible to xts or matrix, assumed to contain Open - High - Low - Close - Volume data.}

\item{n}{Number of periods for moving average. Default is 20.}

\item{nSig}{Number of periods for signal line moving average. Default is 9.}

\item{maType}{Either:
\enumerate{
\item A function or a string naming the function to be called.
\item A \emph{list} with the first component like (1) above, and
additional parameters specified as \emph{named} components.
See Examples.
}}

\item{percent}{logical; if \code{TRUE}, the rate of change is calculated
using the \code{ROC} function, otherwise the \code{momentum} function is
used. Default is \code{TRUE}.}

\item{append}{A logical value. If \code{TRUE}, the calculated TRIX and signal line
values will be appended to the \code{OHLCV} input data, ensuring
proper alignment of time - series data. If \code{FALSE}, only the calculated
TRIX and signal line values will be returned. Defaults to \code{FALSE}.}

\item{...}{Other arguments to be passed to the \code{maType} function in
case (1) above.}
}
\value{
If \code{append = FALSE}, an object of the same class as \code{OHLCV}
or a vector (if \code{try.xts} fails) containing the TRIX and signal line values.
If \code{append = TRUE}, an object of the same class as \code{OHLCV} with the
calculated TRIX and signal line values appended, maintaining the integrity of the time - series
alignment.
}
\description{
The TRIX indicator calculates the rate of change of a triple exponential
moving average. Developed by Jack K. Hutson.
The TRIX is calculated as follows:\cr 3MA = \code{MA}( \code{MA}(
\code{MA}(\code{price}) ) )\cr \eqn{trix = 100 * [ 3MA(t) / 3MA(t - 1) - 1 ]}
}
\note{
Buy/sell signals are generated when the TRIX crosses above/below zero.
A nine - period EMA of the TRIX is used as a default signal line. Buy/sell
signals are generated when the TRIX crosses above/below the signal line and
is also above/below zero.
}
\examples{
\dontrun{
data(TSLA)
# Using default parameters without appending
trix_result1 <- add_TRIX(TSLA)

# Modifying n and without appending
trix_result2 <- add_TRIX(TSLA, n = 25)

# Using default parameters and appending
trix_result3 <- add_TRIX(TSLA, append = TRUE)

# Modifying n and appending
trix_result4 <- add_TRIX(TSLA, n = 25, append = TRUE)
}
}
\references{
The following site(s) were used to code/document this
indicator:\cr
\url{https://www.fmlabs.com/reference/default.htm?url=TRIX.htm}\cr
\url{https://www.metastock.com/Customer/Resources/TAAZ/?p=114}\cr
\url{https://www.linnsoft.com/techind/trix - triple - smoothed - exponential - oscillator}\cr
\url{https://school.stockcharts.com/doku.php?id=technical_indicators:trix}\cr
}
\seealso{
See \code{\link{EMA}}, \code{\link{SMA}}, etc. for moving average
options; and note Warning section.
}
\author{
DengYishuo
}
\keyword{ts}
