% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_TSI.R
\name{add_TSI}
\alias{add_TSI}
\title{Calculate True Strength Index (TSI)}
\usage{
add_TSI(OHLCV, r = 13, s = 25, signal_period = 9, append = FALSE, ...)
}
\arguments{
\item{OHLCV}{Object that is coercible to xts or matrix and contains Open - High - Low - Close - Volume data.
The function will extract the closing price from this object for TSI calculation.}

\item{r}{The period for the first exponential moving average (fast EMA), default is 13.
A shorter period makes the fast EMA more responsive to recent price changes,
while a longer period smooths out the data more.}

\item{s}{The period for the second exponential moving average (slow EMA), default is 25.
Similar to \code{r}, it affects the responsiveness and smoothness of the slow EMA.}

\item{signal_period}{The period for the signal line (EMA of TSI), default is 9.
This determines how smooth the signal line is, with a shorter period
making it more responsive to changes in the TSI.}

\item{append}{A logical value. If \code{TRUE}, the calculated TSI and signal line
values will be appended to the input \code{OHLCV} data (if it's an xts object), ensuring
proper alignment of time - series data. If \code{FALSE}, only the calculated
TSI and signal line values will be returned. Defaults to \code{FALSE}.}

\item{...}{Additional arguments passed to the EMA function (e.g., wilder).
These can be used to customize the EMA calculation, such as using Welles Wilder's
EMA formula if \code{wilder = TRUE}.}
}
\value{
If \code{append = FALSE}, an object of the same type as the input (vector or xts) containing two columns:
"tsi" (True Strength Index values) and "signal" (signal line values).
If \code{append = TRUE} and the input \code{OHLCV} is an xts object, an xts object with the calculated
TSI and signal line values appended, maintaining the integrity of the time - series alignment.
If \code{append = TRUE} and the input \code{OHLCV} is not an xts object, an error is thrown.
}
\description{
The True Strength Index (TSI) is a momentum - based oscillator that measures the strength
of price trends. It uses double exponential moving averages (EMAs) of price changes and
their absolute values to calculate an oscillator value. This function computes the TSI
and an optional signal line (an EMA of the TSI) from the closing prices within the OHLCV data.
}
