% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_ultimateOscillator.R
\name{add_ultimateOscillator}
\alias{add_ultimateOscillator}
\title{Calculate The Ultimate Oscillator}
\usage{
add_ultimateOscillator(
  OHLCV,
  n = c(7, 14, 28),
  wts = c(4, 2, 1),
  append = FALSE
)
}
\arguments{
\item{OHLCV}{Object that is coercible to xts or matrix, assumed to contain Open - High - Low - Close - Volume data.}

\item{n}{A vector of the number of periods to use for each average calculation. Default is c(7, 14, 28).}

\item{wts}{The weights applied to each average. Default is c(4, 2, 1).}

\item{append}{A logical value. If \code{TRUE}, the calculated Ultimate Oscillator
values will be appended to the \code{OHLCV} input data, ensuring
proper alignment of time - series data. If \code{FALSE}, only the calculated
Ultimate Oscillator values will be returned. Defaults to \code{FALSE}.}
}
\value{
If \code{append = FALSE}, an object of the same class as \code{OHLCV}
or a matrix (if \code{try.xts} fails) containing the Ultimate Oscillator values.
If \code{append = TRUE}, an object of the same class as \code{OHLCV} with the
calculated Ultimate Oscillator values appended, maintaining the integrity of the time - series
alignment.
}
\description{
The Ultimate Oscillator is a momentum oscillator designed to capture momentum across three
different time frames.
Created by Larry Williams in 1976.
}
\examples{
\dontrun{
data(TSLA)
# Using default parameters without appending
ult_osc_result1 <- add_ultimateOscillator(TSLA)

# Modifying n and without appending
ult_osc_result2 <- add_ultimateOscillator(TSLA, n = c(5, 10, 20))

# Using default parameters and appending
ult_osc_result3 <- add_ultimateOscillator(TSLA, append = TRUE)

# Modifying n and appending
ult_osc_result4 <- add_ultimateOscillator(TSLA, n = c(5, 10, 20), append = TRUE)
}
}
\references{
The following site(s) were used to code/document this
indicator:\cr
\url{https://school.stockcharts.com/doku.php?id=technical_indicators:ultimate_oscillator}\cr
}
\author{
Ivan Popivanov
}
\keyword{ts}
