% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_volatility.R
\name{add_volatility}
\alias{add_volatility}
\title{Calculate Volatility}
\usage{
add_volatility(
  OHLCV,
  n = 10,
  calc = "close",
  N = 260,
  mean0 = FALSE,
  append = FALSE,
  ...
)
}
\arguments{
\item{OHLCV}{Object that is coercible to xts or matrix and contains
Open-High-Low-Close prices (or only Close prices, if \code{calc="close"}).}

\item{n}{Number of periods for the volatility estimate. Default is 10.}

\item{calc}{The calculation (type) of estimator to use.}

\item{N}{Number of periods per year. Default is 260.}

\item{mean0}{Use a mean of 0 rather than the sample mean. Default is FALSE.}

\item{append}{A logical value. If \code{TRUE}, the calculated volatility
values will be appended to the \code{OHLCV} input data, ensuring
proper alignment of time-series data. If \code{FALSE}, only the calculated
volatility values will be returned. Defaults to \code{FALSE}.}

\item{...}{Arguments to be passed to/from other methods.}
}
\value{
If \code{append = FALSE}, an object of the same class as \code{OHLCV}
or a vector (if \code{try.xts} fails) containing the chosen volatility estimator values.
If \code{append = TRUE}, an object of the same class as \code{OHLCV} with the
calculated volatility values appended, maintaining the integrity of the time-series
alignment.
}
\description{
Selected volatility estimators/indicators; various authors.
}
\examples{
\dontrun{
data(TSLA)
ohlcv <- TSLA[, c("Open", "High", "Low", "Close")]
# Using default parameters without appending
vClose <- add_volatility(ohlcv, calc = "close")
# Modifying n and without appending
vClose0 <- add_volatility(ohlcv, calc = "close", n = 15)
# Using default parameters and appending
vGK <- add_volatility(ohlcv, calc = "garman", append = TRUE)
# Modifying N and appending
vParkinson <- add_volatility(ohlcv, calc = "parkinson", N = 250, append = TRUE)
# Using default parameters and appending for Rogers-Satchell
vRS <- add_volatility(ohlcv, calc = "rogers", append = TRUE)
}
}
\references{
The following sites were used to code/document these
indicators. All were created by Thijs van den Berg under the GNU Free
Documentation License and were retrieved on 2008-04-20. The original
links are dead, but can be accessed via internet archives.\cr
\cr Close-to-Close Volatility (\code{calc="close"}):\cr
\url{https://web.archive.org/web/20100421083157/http://www.sitmo.com/eq/172}\cr
\cr OHLC Volatility: Garman Klass (\code{calc="garman.klass"}):\cr
\url{https://web.archive.org/web/20100326172550/http://www.sitmo.com/eq/402}\cr
\cr High-Low Volatility: Parkinson (\code{calc="parkinson"}):\cr
\url{https://web.archive.org/web/20100328195855/http://www.sitmo.com/eq/173}\cr
\cr OHLC Volatility: Rogers Satchell (\code{calc="rogers.satchell"}):\cr
\url{https://web.archive.org/web/20091002233833/http://www.sitmo.com/eq/414}\cr
\cr OHLC Volatility: Garman Klass - Yang Zhang (\code{calc="gk.yz"}):\cr
\url{https://web.archive.org/web/20100326215050/http://www.sitmo.com/eq/409}\cr
\cr OHLC Volatility: Yang Zhang (\code{calc="yang.zhang"}):\cr
\url{https://web.archive.org/web/20100326215050/http://www.sitmo.com/eq/409}\cr
}
\seealso{
See \code{\link{TR}} and \code{\link{chaikinVolatility}} for other
volatility measures.
}
\author{
DengYishuo
}
\keyword{ts}
