% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_williamsAD.R
\name{add_williamsAD}
\alias{add_williamsAD}
\title{Calculate Williams Accumulation / Distribution}
\usage{
add_williamsAD(OHLCV, append = FALSE)
}
\arguments{
\item{OHLCV}{Object that is coercible to xts or matrix, assumed to contain Open - High - Low - Close - Volume data.}

\item{append}{A logical value. If \code{TRUE}, the calculated Williams AD values will be appended to the \code{OHLCV} input data, ensuring
proper alignment of time - series data. If \code{FALSE}, only the calculated
Williams AD values will be returned. Defaults to \code{FALSE}.}
}
\value{
If \code{append = FALSE}, an object of the same class as \code{OHLCV}
or a vector (if \code{try.xts} fails) containing the accumulation / distribution values.
If \code{append = TRUE}, an object of the same class as \code{OHLCV} with the
calculated Williams AD values appended, maintaining the integrity of the time - series
alignment.
}
\description{
The Williams Accumulation / Distribution (AD) line is a measure of market
momentum. Developed by Larry Williams.
The Williams AD line differs from OBV and chaikinAD in that it doesn't take
volume into account.
}
\note{
The Accumulation/Distribution Line is interpreted by looking for a
divergence in the direction of the indicator relative to price.
}
\examples{
\dontrun{
data(TSLA)
# Using default parameters without appending
ad_result1 <- add_williamsAD(TSLA)

# Using default parameters and appending
ad_result2 <- add_williamsAD(TSLA, append = TRUE)
}
}
\references{
The following site(s) were used to code/document this
indicator:\cr
\url{https://www.fmlabs.com/reference/WilliamsAD.htm}\cr
\url{https://www.metastock.com/Customer/Resources/TAAZ/?p=125}\cr
}
\seealso{
See \code{\link{OBV}}, \code{\link{chaikinAD}}, and
\code{\link{ATR}}.
}
\author{
DengYishuo
}
\keyword{ts}
