% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_WPR.R
\name{add_WPR}
\alias{add_WPR}
\title{Calculate Williams \%R}
\usage{
add_WPR(OHLCV, n = 14, scale = FALSE, append = FALSE)
}
\arguments{
\item{OHLCV}{Object coercible to xts/matrix, assumed to contain Open - High - Low - Close - Volume data.}

\item{n}{Lookback period (default 14).}

\item{scale}{Scale to \code{[0, -100]} range? (default FALSE).}

\item{append}{A logical value. If \code{TRUE}, the calculated Williams \%R values will be appended to the \code{OHLCV} input data, ensuring
proper alignment of time - series data. If \code{FALSE}, only the calculated
Williams \%R values will be returned. Defaults to \code{FALSE}.}
}
\value{
If \code{append = FALSE}, an object matching \code{OHLCV} class containing Williams \%R values.
If \code{append = TRUE}, an object of the same class as \code{OHLCV} with the
calculated Williams \%R values appended, maintaining the integrity of the time - series
alignment.
}
\description{
Measures closing price's position relative to price range over last \code{n} periods.
}
\details{
\deqn{WPR = \frac{\mathit{highest\ high} - close}{\mathit{highest\ high} - \mathit{lowest\ low}}}
Equivalent to \eqn{1 - fastK} from stochastic oscillator.
}
\note{
Requires \code{runMax}/\code{runMin} from eTTR/quantmod.
}
\examples{
\dontrun{
data(TSLA)
ohlcv <- TSLA[, c("Open", "High", "Low", "Close")]
# Using default parameters without appending
wpr_result1 <- add_WPR(ohlcv)

# Using default parameters and appending
wpr_result2 <- add_WPR(ohlcv, append = TRUE)

# Scaling and without appending
wpr_scaled_result1 <- add_WPR(ohlcv, scale = TRUE)

# Scaling and appending
wpr_scaled_result2 <- add_WPR(ohlcv, scale = TRUE, append = TRUE)
}
}
\references{
\itemize{
\item \url{https://www.fmlabs.com/reference/WilliamsR.htm}
\item \url{https://school.stockcharts.com/doku.php?id=technical_indicators:williams_r}
}
}
\seealso{
Related function: \code{\link[=stoch]{stoch()}}
}
\keyword{ts}
