% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_ZigZag.R
\name{add_ZigZag}
\alias{add_ZigZag}
\alias{add_zigzag}
\title{Calculate Zig Zag}
\usage{
add_ZigZag(
  OHLCV,
  change = 10,
  percent = TRUE,
  retrace = FALSE,
  lastExtreme = TRUE,
  append = FALSE
)
}
\arguments{
\item{OHLCV}{Object that is coercible to xts or matrix, assumed to contain
Open - High - Low - Close - Volume data.}

\item{change}{Minimum price movement, either in dollars or percent (see
\code{percent}).}

\item{percent}{Use percentage or dollar change?}

\item{retrace}{Is \code{change} a retracement of the previous move, or an
absolute change from peak to trough?}

\item{lastExtreme}{If the extreme price is the same over multiple periods,
should the extreme price be the first or last observation?}

\item{append}{A logical value. If \code{TRUE}, the calculated Zig Zag values
will be appended to the \code{OHLCV} input data, ensuring
proper alignment of time - series data. If \code{FALSE}, only the calculated
Zig Zag values will be returned. Defaults to \code{FALSE}.}
}
\value{
If \code{append = FALSE}, an object of the same class as \code{OHLCV}
or a vector (if \code{try.xts} fails) containing the Zig Zag indicator.
If \code{append = TRUE}, an object of the same class as \code{OHLCV} with the
calculated Zig Zag values appended, maintaining the integrity of the time - series
alignment.
}
\description{
Zig Zag higlights trends by removing price changes smaller than \code{change}
and interpolating lines between the extreme points.

The Zig Zag is non - predictive. The purpose of the Zig Zag is filter noise
and make chart patterns clearer. It's more a visual tool than an indicator.
}
\note{
If High - Low prices are given, the function calculates the max/min using
the high/low prices. Otherwise the function calculates the max/min of the
single series.
}
\section{Warning}{
 The last value of the ZigZag indicator is unstable (i.e.
unknown) until the turning point actually occurs. Therefore this indicator
isn't well - suited for use for systematic trading strategies.
}

\examples{
\dontrun{
# Using default parameters without appending
data(TSLA)
zz_result1 <- add_ZigZag(TSLA, change = 20)

# Using default parameters and appending
zz_result2 <- add_ZigZag(TSLA, change = 20, append = TRUE)

# Changing parameters and without appending
zz_result3 <- add_ZigZag(TSLA, change = 10, percent = FALSE, retrace = TRUE)

# Changing parameters and appending
zz_result4 <- add_ZigZag(TSLA, change = 10, percent = FALSE, retrace = TRUE, append = TRUE)
}
}
\references{
The following site(s) were used to code/document this
indicator:\cr
\url{https://www.fmlabs.com/reference/default.htm?url=ZigZag.htm}\cr
\url{https://www.linnsoft.com/techind/zig - zag - indicator - zig - zzo}\cr
\url{https://www.linnsoft.com/techind/zig - zag - oscillator - indicator - zzo}\cr
\url{https://www.metastock.com/Customer/Resources/TAAZ/?p=127}\cr
\url{https://school.stockcharts.com/doku.php?id=technical_indicators:zigzag}\cr
}
\author{
DengYishuo
}
\keyword{ts}
