% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_ZLEMA.R
\name{add_ZLEMA}
\alias{add_ZLEMA}
\title{Calculate Zero - Lag Exponential Moving Average (ZLEMA)}
\usage{
add_ZLEMA(OHLCV, n = 10, ratio = NULL, append = FALSE)
}
\arguments{
\item{OHLCV}{Object that is coercible to xts or matrix and contains Open - High - Low - Close - Volume data.
The function will extract the closing price from this object for ZLEMA calculation.}

\item{n}{Number of periods to average over. This parameter affects the smoothness and
responsiveness of the ZLEMA. A larger \code{n} will result in a smoother ZLEMA,
but it may be less responsive to recent price changes.}

\item{ratio}{A smoothing/decay ratio. This ratio determines how much weight is given
to recent observations in the ZLEMA calculation. If specified, it overrides the
default calculation based on \code{n}.}

\item{append}{A logical value. If \code{TRUE}, the calculated ZLEMA values will be appended to the \code{OHLCV} input data,
ensuring proper alignment of time - series data. If \code{FALSE}, only the calculated
ZLEMA values will be returned. Defaults to \code{FALSE}.}
}
\value{
If \code{append} is \code{FALSE}, an object of the same class as \code{OHLCV} (or a vector if \code{try.xts} fails)
containing the ZLEMA values.
If \code{append} is \code{TRUE}, an object of the same class as \code{OHLCV} with the calculated ZLEMA values appended,
maintaining the integrity of the time - series alignment.
}
\description{
Calculate an exponential moving average with reduced lag. The Zero - Lag Exponential Moving
Average (ZLEMA) is designed to provide a more timely representation of price trends
by reducing the lag typically associated with traditional exponential moving averages.
}
\examples{
\dontrun{
data(TSLA)
# Using default parameters without appending
zlema_result1 <- add_ZLEMA(TSLA)

# Using default parameters and appending
zlema_result2 <- add_ZLEMA(TSLA, append = TRUE)

# Changing n and without appending
zlema_result3 <- add_ZLEMA(TSLA, n = 15)

# Changing n and appending
zlema_result4 <- add_ZLEMA(TSLA, n = 15, append = TRUE)
}
}
\keyword{ts}
