% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/momentum.R
\name{momentum}
\alias{momentum}
\title{Momentum}
\usage{
momentum(x, n = 1, na.pad = TRUE)
}
\arguments{
\item{x}{Price, volume, etc. series that is coercible to xts or matrix.}

\item{n}{Number of periods to use.}

\item{na.pad}{Should periods prior to \code{n} be appended? Default is \code{TRUE}.}
}
\value{
A object of the same class as \code{x} or a vector (if \code{try.xts}
fails) containing the momentum values.
}
\description{
Calculate the momentum of a series over \code{n} periods.

The momentum indicator provides the difference between the current value
and the value \code{n} periods ago.
}
\examples{
data(TSLA)
mom <- momentum(TSLA[, "Close"])
}
\author{
DengYishuo
}
\keyword{ts}
