% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OBV.R
\name{OBV}
\alias{OBV}
\title{On Balance Volume (OBV)}
\usage{
OBV(price, volume)
}
\arguments{
\item{price}{Price series that is coercible to xts or matrix.}

\item{volume}{Volume series that is coercible to xts or matrix,
corresponding to the price object.}
}
\value{
A object of the same class as \code{price} and \code{volume} or a
vector (if \code{try.xts} fails) containing the OBV values.
}
\description{
On Balance Volume (OBV) is a momentum-based technical indicator that measures
the cumulative flow of volume to determine the strength of a price trend.
It uses volume flow to predict changes in stock prices by adding volume on
up days and subtracting volume on down days.
}
\note{
OBV is typically used to confirm price trends or identify potential
divergences between volume flow and price movements.
}
\examples{
data(TSLA)
obv <- OBV(TSLA[, "Close"], ttrc[, "Volume"])
}
\references{
\itemize{
\item https://www.fmlabs.com/reference/OBV.htm
\item https://www.metastock.com/Customer/Resources/TAAZ/?p=82
\item https://www.linnsoft.com/techind/balance-open-interest
\item https://school.stockcharts.com/doku.php?id=technical_indicators:on_balance_volume_obv
}
}
\seealso{
See \code{\link{chaikinAD}} for a similar volume-based indicator.
}
\author{
DengYishuo
}
\keyword{indicator}
\keyword{momentum}
\keyword{ts}
\keyword{volume}
