% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eTTR-package.R
\docType{data}
\name{TSLA}
\alias{TSLA}
\title{Tesla Inc. Stock Price Data}
\format{
A data frame with 5 columns:
\describe{
  \item{Open}{Opening price}
  \item{High}{Highest price}
  \item{Low}{Lowest price}
  \item{Close}{Closing price}
  \item{Volume}{Trading volume}
}
}
\source{
Yahoo Finance
}
\description{
Daily stock prices for Tesla Inc. (TSLA) from 2010 to present.
}
