% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VHF.R
\name{VHF}
\alias{VHF}
\title{Vertical Horizontal Filter (Optimized)}
\usage{
VHF(price, n = 28, na.rm = TRUE, inf.replace = NA, zero.replace = NA)
}
\arguments{
\item{price}{Object that is coercible to xts or matrix and contains a Close
price series, or a High-Low-Close price series.}

\item{n}{Number of periods to use. Default is 28.}

\item{na.rm}{Logical. Should NA values be removed during calculations? Default is TRUE.}

\item{inf.replace}{Numeric value to replace INF/-INF values. Default is NA.}

\item{zero.replace}{Numeric value to replace division by zero results. Default is NA.}
}
\value{
A object of the same class as \code{price} or a vector (if
\code{try.xts} fails) containing the VHF values with NA/INF values handled.
}
\description{
The Vertical Horizontal Filter (VHF) attempts to identify starting and ending
trends. This optimized version includes safeguards against NA and INF values.
}
\note{
If Close prices are given, the function calculates the max/min using
only those prices. If HLC prices are given, the function uses high/low prices.
}
\examples{
data(TSLA)
vhf_close <- VHF(TSLA[, "Close"], na.rm = TRUE)
vhf_hlc <- VHF(TSLA[, c("High", "Low", "Close")], inf.replace = 10)
}
\references{
\url{https://www.metastock.com/Customer/Resources/TAAZ/?p=119}
}
\seealso{
See \code{\link{aroon}}, \code{\link{CCI}}, \code{\link{ADX}},
\code{\link{TDI}}, \code{\link{GMMA}} for other trend indicators.
}
\author{
Deng Yishuo
}
\keyword{ts}
