% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wilderSum.R
\name{wilderSum}
\alias{wilderSum}
\title{Welles Wilder Style Sum (wilderSum)}
\usage{
wilderSum(x, n = 10)
}
\arguments{
\item{x}{Object coercible to xts or matrix.}

\item{n}{Number of periods in the window (1 <= n <= nrow(x)).}
}
\value{
An object of the same class as \code{x} with Wilder-style sums.
}
\description{
Calculate a weighted sum using Welles Wilder's method.
}
\examples{
data(ttrc)
wilder_14 <- wilderSum(ttrc[, "High"] - ttrc[, "Low"], 14)
head(wilder_14)
}
\keyword{internal}
\keyword{ts}
